### 2019-12-18 #1962 @whedon
[PRE REVIEW]: LiberTEM: Software platform for scalable multidimensional data processing in transmission electron microscopy
**Submitting author:** @uellue (<a href="http://orcid.org/0000-0001-6635-9567">Dieter Weber</a>)
**Repository:** <a href="https://github.com/LiberTEM/LiberTEM" target ="_blank">https://github.com/LiberTEM/LiberTEM</a>
**Version:** v0.3.0
