### 2019-12-18 f3d6b4f
HDF5 tests: create alias of time.time for mocking
