+++ tests/io/test_hdf5.py
@@ -183,10 +183,9 @@ def test_timeout_1(hdf5, lt_ctx):
         print(diags)
 
 
-@pytest.mark.skip(reason="mocking time.time is dangerous in multi-threaded environments")
 def test_timeout_2(hdf5, lt_ctx):
     print(threading.enumerate())
-    with mock.patch('time.time', side_effect=[1, 30]):
+    with mock.patch('libertem.io.dataset.hdf5.current_time', side_effect=[1, 30]):
         params = H5DataSet.detect_params(hdf5.filename, executor=lt_ctx.executor)
         assert list(params.keys()) == ["path"]
 
@@ -196,7 +195,7 @@ def test_timeout_2(hdf5, lt_ctx):
     ds = ds.initialize(lt_ctx.executor)
 
     print(threading.enumerate())
-    with mock.patch('time.time', side_effect=[30, 60]):
+    with mock.patch('libertem.io.dataset.hdf5.current_time', side_effect=[30, 60]):
         diags = ds.diagnostics
         print(diags)
 
