### 2019-12-19 #1962 @uellue
> [ point_right Check article proof page_facing_up point_left ](https://github.com/openjournals/joss-papers/blob/joss.01962/joss.01962/10.21105.joss.01962.pdf)

The proof looks good!
