### 2019-12-20 #1962 @labarba
> Hi, the invalid DOI is of a book that is in print and I am referring to a chapter that I co-authored: https://www.worldscientific.com/worldscibooks/10.1142/11389
> How should we deal with this? The publication date of the book was pushed recently from Jan 2020 to March 2020, by the way.

Our bot's suggestions on DOIs are best-effort. Do what you think is best!
