### 2019-12-20 #1962 @labarba
... one more thing: is there any chance you could deposit your chapter on a pre-print server?
