### 2019-12-20 #1962 @labarba
👋  @majensen — could you handle this submission as editor? I cannot identify any editor in the general domain of this software, but it is aimed at processing large-scale binary data sets, which I think is in scope for you?
