### 2019-12-20 #1962 @labarba
@uellue : in the paper, you have several points where you say that something is in the docs, and embed a long URL to the position in the docs. This looks rather ugly, and it's unnecessary. Can you clean that up a bit? You can have just one link to the docs, and so shorten that section of the text quite a bit by just describing briefly: "here are our docs, and they contain this, that, and more."
