### 2019-12-20 #1962 @uellue
> @uellue : (...) You can have just one link to the docs, and so shorten that section of the text quite a bit by just describing briefly: "here are our docs, and they contain this, that, and more."

Will do, thx for the feedback! :-)
