### 2019-12-20 66b3d56
Consolidate documentation links
Feedback by @labarba https://github.com/openjournals/joss-reviews/issues/1962#issuecomment-567965914

Thx! :-)
