+++ docs/publications/joss/paper.bib
@@ -82,15 +82,6 @@
   publisher = {Zenodo},
 }
 
-@Misc{Clausen2019,
-  author    = {Clausen, Alexander and Weber, Dieter and Ruzaeva, Karina and Caron, Jan and Chandra, Rahul and Migunov, Vadim and Nord, Magnus and Ophus, Colin and Peter, Simon and van Schyndel, Jay and Shin, Jaeweon and Müller-Caspary, Knut and Dunin-Borkowski, Rafal},
-  title     = {LiberTEM/LiberTEM: 0.2.2},
-  year      = {2019},
-  doi       = {10.5281/ZENODO.1477847},
-  keywords  = {STEM, TEM, pixelated STEM, 4D STEM, high-throughput, electron microscopy},
-  publisher = {Zenodo},
-}
-
 @InBook{Weber2020,
   chapter   = {Next-Generation Information Technology Systems for Fast High-Resolution Detectors in Electron Microscopy},
   title     = {Handbook on {Big} {Data} and machine learning in the physical sciences},
@@ -115,6 +106,15 @@
   month     = {may},
   doi       = {10.1017/s1431927619000497},
   publisher = {Cambridge University Press ({CUP})},
+}
+
+@Misc{Clausen2019a,
+  author    = {Clausen, Alexander and Weber, Dieter and Ruzaeva, Karina and Migunov, Vadim and Caron, Jan and Chandra, Rahul and Nord, Magnus and Ophus, Colin and Peter, Simon and Schyndel van, Jay and Shin, Jaeweon and Müller-Caspary, Knut and Dunin-Borkowski, Rafal},
+  title     = {LiberTEM/LiberTEM: 0.3.0},
+  year      = {2019},
+  doi       = {10.5281/ZENODO.3572855},
+  keywords  = {STEM, TEM, pixelated STEM, 4D STEM, high-throughput, electron microscopy},
+  publisher = {Zenodo},
 }
 
 @Comment{jabref-meta: databaseType:bibtex;}
