+++ docs/publications/joss/paper.bib
@@ -47,7 +47,7 @@
 
 @Article{Zaharia2016,
   author    = {Matei Zaharia and Michael J. Franklin and Ali Ghodsi and Joseph Gonzalez and Scott Shenker and Ion Stoica and Reynold S. Xin and Patrick Wendell and Tathagata Das and Michael Armbrust and Ankur Dave and Xiangrui Meng and Josh Rosen and Shivaram Venkataraman},
-  title     = {Apache Spark},
+  title     = {Apache {Spark}},
   journal   = {Communications of the {ACM}},
   year      = {2016},
   volume    = {59},
