### 2019-12-21 #1962 @majensen
Hey @cr458 - would you be able to review [this submission](https://github.com/openjournals/joss-papers/blob/joss.01962/joss.01962/10.21105.joss.01962.pdf) to [JOSS](https://joss.theoj.org)? Thanks for considering! --@majensen
