### 2020-01-06 1ad0c22
DaskJobExecutor.run_each_host: need to pass pure=False
Otherwise, all Futures map to the same scheduler key and thus
only one of them will be executed, even though they have different
workers constraints.
