### 2020-01-06 60c29fb
Tests for distributed behavior
This uses docker and docker-compose to spin up a scheduler and two
workers.
