+++ packaging/docker/update_reqs.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+pip-compile ../../setup.py -Uo requirements.txt
