+++ tox.ini
@@ -16,6 +16,8 @@ setenv=
     # leads to collisions between the local source tree and the installed package when running tests.
     # See https://github.com/pytest-dev/pytest/issues/2042
     PY_IGNORE_IMPORTMISMATCH=1
+passenv=
+    DASK_SCHEDULER_ADDRESS
 
 [testenv:flake8]
 changedir={toxinidir}
