+++ setup.cfg
@@ -37,7 +37,8 @@ exclude_lines =
 
 [tool:pytest]
 norecursedirs = .git .tox *.egg-info build node_modules TOXENV benchmarks prototypes
-addopts = --strict -m "not slow"
+addopts = --strict -m "not slow" -m "not dist"
 markers =
     slow: mark a test as slow, i.e. takes a couple of seconds to run
+    dist: tests that exercise the distributed parts of libertem
     functional: mark a test as functional, testing the whole libertem stack
