+++ packaging/docker/requirements.txt
@@ -1,0 +1,50 @@
+#
+# This file is autogenerated by pip-compile
+# To update, run:
+#
+#    pip-compile --output-file=requirements.txt ../../setup.py
+#
+attrs==19.3.0             # via jsonschema
+click==7.0
+cloudpickle==1.2.2
+cycler==0.10.0            # via matplotlib
+dask==2.9.1               # via distributed
+decorator==4.4.1          # via networkx
+distributed==2.9.1
+h5py-cache==1.0           # via ncempy
+h5py==2.10.0
+heapdict==1.0.1           # via zict
+imageio==2.6.1            # via scikit-image
+importlib-metadata==1.3.0  # via jsonschema
+joblib==0.14.1            # via scikit-learn
+jsonschema==3.2.0
+kiwisolver==1.1.0         # via matplotlib
+llvmlite==0.31.0          # via numba
+matplotlib==3.1.2
+more-itertools==8.0.2     # via zipp
+msgpack==0.6.2            # via distributed
+ncempy==1.5.0
+networkx==2.4             # via scikit-image
+numba==0.47.0
+numpy==1.18.0
+pillow==7.0.0
+psutil==5.6.7
+pyparsing==2.4.6          # via matplotlib
+pyrsistent==0.15.6        # via jsonschema
+python-dateutil==2.8.1    # via matplotlib
+pywavelets==1.1.1         # via scikit-image
+pyyaml==5.2               # via distributed
+scikit-image==0.16.2
+scikit-learn==0.22.1
+scipy==1.4.1
+six==1.13.0               # via cycler, h5py, jsonschema, pyrsistent, python-dateutil
+sortedcontainers==2.1.0   # via distributed
+sparse==0.8.0
+tblib==1.6.0              # via distributed
+toolz==0.10.0             # via distributed
+tornado==6.0.3
+zict==1.0.0               # via distributed
+zipp==0.6.0               # via importlib-metadata
+
+# The following packages are considered to be unsafe in a requirements file:
+# setuptools
