### 2020-01-06 4b14004
Actually set DASK_SCHEDULER_ADDRESS for dist tests job
