### 2020-01-06 7932f05
Specify python version for dist travis job
