### 2020-01-06 8923062
Update packaging/docker/requirements.txt in CI
