+++ .travis.yml
@@ -47,6 +47,9 @@ jobs:
       dist: xenial
       services:
         - docker
+      install:
+        - pip install tox pip-tools
+        - cd packaging/docker/ && ./update_reqs.sh
       script:
         - cd packaging/docker/ && docker-compose up --build -d
         - tox -e py37 -- tests/ -m dist
