### 2020-01-06 ff27702
Update documentation
 * Re-enable changelog globbing mechanism
 * Add changelog entry for executor bugfix
 * Add documentation for running distributed tests
