### 2020-01-06 99185cf
Update documentation
 * Re-enable changelog globbing mechanism
 * Add changelog entry for executor bugfix
 * Add documentation for running distributed tests
