+++ docs/source/changelog/bugfix/run_each_host_nonpure.rst
@@ -1,0 +1,4 @@
+Fix DaskJobExecutor.run_each_host
+=================================
+
+ * Need to pass :code:`pure=False` to ensure multiple runs of the function (:pr:`528`).
