### 2020-01-06 ef62793
Run `cd` in subshell to keep cwd intact
