### 2020-01-06 22d6b87
Update docs: add the commands for running dist tests
