+++ src/libertem/io/dataset/frms6.py
@@ -97,6 +97,13 @@ def _get_base_filename(_path):
 
 
 def _read_hdr(fname):
+    if not os.path.exists(fname):
+        raise DataSetException(
+            "Could not find .hdr file %s" % (
+                fname,
+            )
+        )
+
     config = configparser.ConfigParser()
     config.read(fname)
     parsed = {}
@@ -130,9 +137,11 @@ def _read_hdr(fname):
 class FRMS6File(File3D):
     def __init__(self, path, start_idx=None, hdr_info=None):
         self._path = path
-        self._header = None
         self._hdr_info = hdr_info
         self._start_idx = start_idx
+        self._header = None
+        self._size = None
+        self._preread_info()
         super().__init__()
 
     @property
@@ -159,6 +168,10 @@ class FRMS6File(File3D):
     def global_header(self):
         return self._hdr_info
 
+    def _preread_info(self):
+        self._header = self.header
+        self._size = os.stat(self._path).st_size
+
     def check_valid(self):
         if not (self.header['header_size'] == 1024
                 and self.header['frame_header_size'] == 64
@@ -173,11 +186,11 @@ class FRMS6File(File3D):
             return self.header['num_frames']
         # older FRMS6 files don't contain the number of frames in the header,
         # so calculate from filesize:
-        size = os.stat(self._path).st_size
         header = self.header
         w, h = header['width'], header['height']
         bytes_per_frame = w * h * 2
-        num = (size - 1024)
+        assert self._size is not None
+        num = (self._size - 1024)
         denum = (bytes_per_frame + 64)
         res = num // denum
         if num % denum != 0:
@@ -335,6 +348,9 @@ class FRMS6DataSet(DataSet):
         self._dark_frame = None
         self._enable_offset_correction = enable_offset_correction
         self._meta = None
+        self._filenames = None
+        self._hdr_info = None
+        self._fileset = None
         if dest_dtype is not None:
             warnings.warn(
                 "dest_dtype is now handled per `get_tiles` call, and ignored here",
@@ -346,6 +362,8 @@ class FRMS6DataSet(DataSet):
         return self._meta.shape
 
     def _do_initialize(self):
+        self._filenames = list(sorted(glob.glob(self._pattern())))
+        self._hdr_info = self._read_hdr_info()
         first_file = next(self._get_signal_files())
         header = first_file.header
         raw_frame_size = header['height'], header['width']
@@ -370,6 +388,12 @@ class FRMS6DataSet(DataSet):
             os.stat(path).st_size
             for path in self._files()
         )
+        self._fileset = FRMS6FileSet(
+            files=list(self._get_signal_files()),
+            meta=self._meta,
+            dark_frame=self._dark_frame,
+            gain_map=self._gain_map,
+        )
         return self
 
     def initialize(self, executor):
@@ -388,9 +412,12 @@ class FRMS6DataSet(DataSet):
     def get_msg_converter(cls):
         return FRMS6DatasetParams
 
-    def _get_hdr_info(self):
+    def _read_hdr_info(self):
         hdr_filename = "%s.hdr" % _get_base_filename(self._path)
         return _read_hdr(hdr_filename)
+
+    def _get_hdr_info(self):
+        return self._hdr_info
 
     def _get_dark_frame(self):
         if not self._enable_offset_correction:
@@ -436,12 +463,7 @@ class FRMS6DataSet(DataSet):
             yield f
 
     def _get_fileset(self):
-        return FRMS6FileSet(
-            files=list(self._get_signal_files()),
-            meta=self._meta,
-            dark_frame=self._dark_frame,
-            gain_map=self._gain_map,
-        )
+        return self._fileset
 
     def _pattern(self):
         path, ext = os.path.splitext(self._path)
[…]
