+++ tests/io/test_frms6.py
@@ -22,6 +22,15 @@ def default_frms6(lt_ctx):
     return ds
 
 
+@pytest.fixture
+def dist_frms6(dist_ctx):
+    path = "/data/frms6/C16_15_24_151203_019.hdr"
+    ds = FRMS6DataSet(path=path)
+    ds = ds.initialize(dist_ctx.executor)
+    assert ds._fileset is not None
+    return ds
+
+
 def test_simple_open(default_frms6):
     assert tuple(default_frms6.shape) == (256, 256, 264, 264)
 
@@ -70,6 +79,18 @@ def test_pickle_is_small(default_frms6):
 
 def test_cache_key_json_serializable(default_frms6):
     json.dumps(default_frms6.get_cache_key())
+
+
+@pytest.mark.dist
+def test_dist_process(dist_frms6, dist_ctx):
+    roi = {
+        "shape": "disk",
+        "cx": 5,
+        "cy": 6,
+        "r": 7,
+    }
+    analysis = SumAnalysis(dataset=dist_frms6, parameters={"roi": roi})
+    dist_ctx.run(analysis)
 
 
 # TODO: gain map tests
