### 2020-01-06 4fa8a30
Add changelog entry
