### 2020-01-06 22074ee
test_holography: use tuple for slicing, fixes #529
