### 2020-01-07 #1962 @majensen
@alvarolopez - feliz año nuevo! Would you be willing to review [this paper](#1962) for [JOSS](https://joss.theoj.org)? Information on reviewing can be found [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
Thanks for your consideration! --@majensen
