### 2020-01-07 b0776e6
Remove HDFS support, fixes #38
