+++ src/libertem/io/dataset/__init__.py
@@ -3,7 +3,6 @@ from libertem.io.dataset.base import DataSetException
 
 
 filetypes = {
-    "hdfs": "libertem.io.dataset.hdfs.BinaryHDFSDataSet",
     "hdf5": "libertem.io.dataset.hdf5.H5DataSet",
     "raw": "libertem.io.dataset.raw.RawFileDataSet",
     "mib": "libertem.io.dataset.mib.MIBDataSet",
