### 2020-01-07 f5b7c71
Remove HDFS support, fixes #38
