### 2020-01-07 26ed430
Update documentation for HDFS removal
