### 2020-01-07 13af3a7
Update documentation for HDFS removal
