### 2020-01-07 3d71dbc
Update PR number in changelog
