### 2020-01-07 7132751
Add codeclimate configuration
