### 2020-01-07 5763bd4
codeclimate.yml: need to state version at top
