+++ .codeclimate.yml
@@ -1,3 +1,4 @@
+version: "2"
 plugins:
   tslint:
     enabled: true
