### 2020-01-07 3a5a3a3
Add codeclimate badge to README
