### 2020-01-07 178dfe2
update exclude_patterns
The codeclimate yaml config file overrides the GUI config completely, so
we lost the exclude patterns that were only configured in the GUI.
