+++ .codeclimate.yml
@@ -15,3 +15,19 @@ plugins:
       threshold: "B"
   bandit:
     enabled: true
+exclude_patterns:
+- "config/"
+- "db/"
+- "dist/"
+- "features/"
+- "**/node_modules/"
+- "script/"
+- "**/spec/"
+- "**/test/"
+- "**/tests/"
+- "Tests/"
+- "**/vendor/"
+- "**/*_test.go"
+- "**/*.d.ts"
+- "benchmarks/"
+- "prototypes/"
