### 2020-01-07 1714e4f
Fix tslint errors
