### 2020-01-07 0570e1d
Fix tslint errors
