### 2020-01-07 a58b065
Reduce complexity of CachedPartition.get_tiles
