### 2020-01-07 1e84a75
Disable ROI selector if the current mode is PICK
Fixes #511
