### 2020-01-07 f0bbd3d
Disable ROI selector if the current mode is PICK
Fixes #511
