### 2020-01-08 e24365e
Record coverage for numba njit functions
This is done by marking tests that exercise jitted functions with
the marker `with_numba`, which are then run with NUMBA_DISABLE_JIT=1.
Note that some numba functions are not supported in Python-mode, notably
to_fixed_tuple, so we can't get coverage for those functions (yet).

Fixes #474.
