+++ .travis.yml
@@ -24,6 +24,11 @@ jobs:
       dist: xenial
       sudo: true
     - stage: test
+      env: TOXENV=numba_coverage
+      python: 3.7
+      dist: xenial
+      sudo: true
+    - stage: test
       env: TOXENV=flake8
       python: 3.7
       dist: xenial
