+++ tox.ini
@@ -19,6 +19,12 @@ setenv=
 passenv=
     DASK_SCHEDULER_ADDRESS
 
+[testenv:numba_coverage]
+commands=
+    pytest --durations=5 --cov=libertem --cov-report=term --cov-report=html --cov-report=xml -m with_numba {posargs:tests/}
+setenv=
+    NUMBA_DISABLE_JIT=1
+
 [testenv:flake8]
 changedir={toxinidir}
 deps=
