+++ setup.cfg
@@ -42,3 +42,4 @@ markers =
     slow: mark a test as slow, i.e. takes a couple of seconds to run
     dist: tests that exercise the distributed parts of libertem
     functional: mark a test as functional, testing the whole libertem stack
+    with_numba: mark tests that exercise a numba function so we can re-run for coverage
