### 2020-01-08 17d15c6
Allow dismissing error messages via escape key, closes #437
