### 2020-01-08 afdedae
Allow dismissing error messages via escape key, closes #437
