+++ src/libertem/web/client/static/js/main.aa6cb99b.chunk.js
@@ -1,2 +1,0 @@
-(this["webpackJsonplibertem-client"]=this["webpackJsonplibertem-client"]||[]).push([[0],{231:function(e,t,n){e.exports={default:"ResultImage_default__1X7FO",fallback:"ResultImage_fallback__kO3cw"}},371:function(e,t,n){e.exports=n.p+"static/media/LiberTEM logo-medium.7999e0dc.png"},383:function(e,t,n){e.exports=n(671)},671:function(e,t,n){"use strict";n.r(t);n(384),n(396),n(406);var a,r=n(0),c=n.n(r),i=n(61),o=n(12),l=n(53),s=n(89),u=n(32),d=n(33),m=n(35),p=n(34),E=n(36),f=(n(465),n(684)),h=n(686),b=n(66),g=n(688),O=n(695),y=n(689),v=Object(o.b)((function(e){return{version:e.config.version,revision:e.config.revision}}))((function(e){var t=e.version,n=e.revision;return r.createElement(r.Fragment,null,r.createElement(O.a,{as:"h3"},"This is LiberTEM version ",t," (revision ",n.slice(0,8),")"),r.createElement(y.a,null,r.createElement(y.a.Item,null,r.createElement(y.a.Icon,{name:"github"}),r.createElement(y.a.Content,null,"Find us on ",r.createElement("a",{href:"https://github.com/LiberTEM/LiberTEM"},"GitHub"))),r.createElement(y.a.Item,null,r.createElement(y.a.Icon,{name:"bug"}),r.createElement(y.a.Content,null,"Found a bug? Got a feature request? Please ",r.createElement("a",{href:"https://github.com/LiberTEM/LiberTEM/issues/new"},"open an issue!"))),r.createElement(y.a.Item,null,r.createElement(y.a.Icon,{name:"legal"}),r.createElement(y.a.Content,null,"LiberTEM is licensed under the ",r.createElement("a",{href:"https://github.com/LiberTEM/LiberTEM/blob/master/LICENSE"},"GNU General Public License v3.0"))),r.createElement(y.a.Item,null,r.createElement(y.a.Icon,{name:"book"}),r.createElement(y.a.Content,null,"Read ",r.createElement("a",{href:"https://libertem.github.io/LiberTEM/"},"the documentation"))),r.createElement(y.a.Item,null,r.createElement(y.a.Icon,{name:"user"}),r.createElement(y.a.Content,null,"Read ",r.createElement("a",{href:"https://libertem.github.io/LiberTEM/acknowledgments.html"},"the acknowledgments"))),r.createElement(y.a.Item,null,r.createElement(y.a.Icon,{name:"gitter"}),r.createElement(y.a.Content,null,"Join our ",r.createElement("a",{href:"https://gitter.im/LiberTEM/Lobby"},"chat on gitter!"))),r.createElement(y.a.Item,null,r.createElement(y.a.Icon,{name:"linkify"}),r.createElement(y.a.Content,null,r.createElement("a",{href:"https://doi.org/10.5281/zenodo.1477847"},r.createElement("img",{src:"https://zenodo.org/badge/DOI/10.5281/zenodo.1477847.svg",alt:"doi.org/10.5281/zenodo.1477847"}))))))})),j=n(690),x=n(696);function S(e,t,n){return void 0===n&&void 0===t?{type:e}:void 0===n?{type:e,payload:t}:{type:e,payload:t,meta:n}}!function(e){e.NOT_CONNECTED="CLUSTER_NOT_CONNECTED",e.CONNECT="CLUSTER_CONNECT",e.CONNECTED="CLUSTER_CONNECTED"}(a||(a={}));var C={notConnected:function(){return S(a.NOT_CONNECTED)},connec
[…]
