### 2020-01-09 #1962 @uellue
> > ... one more thing: is there any chance you could deposit your chapter on a pre-print server?
>
> That's a good point! I'll contact the publisher if they would be fine with that.

Unfortunately, the wouldn't be happy about it.
