### 2020-01-09 663c7d3
Clarify that list items are ORd together
Thx Michael Bussmann!
