### 2020-01-09 6b76f45
Configure codeclimate maintainability rules
 * Increase number of lines per file allowed to 400. This should allow
   us to keep things together that belong together, and not force
   the reader to jump around multiple files to understand some part of
   LiberTEM.
 * Increase number of return statements allowed - redux reducers are an
   example of code that needs needs to be structured like this - for
   example, not having a return statement for each case in the reducer
   means we have to watch out for fall-through into the next case etc.

Refs #541
