+++ .codeclimate.yml
@@ -1,4 +1,11 @@
 version: "2"
+checks:
+  file-lines:
+    config:
+      threshold: 400
+  return-statements:
+    config:
+      threshold: 15
 plugins:
   tslint:
     enabled: true
