### 2020-01-09 #2006 @whedon
[REVIEW]: LiberTEM: Software platform for scalable multidimensional data processing in transmission electron microscopy
**Submitting author:** @uellue (<a href="http://orcid.org/0000-0001-6635-9567">Dieter Weber</a>)
**Repository:** <a href="https://github.com/LiberTEM/LiberTEM" target ="_blank">https://github.com/LiberTEM/LiberTEM</a>
**Version:** 0.5.0
