### 2020-01-13 15e3310
DMDataSet: fix filesize calculation
Filesize calculation needs to run on worker node.

Also, catch IndexError and re-throw as DataSetException
with a more useful message.
