### 2020-01-13 0dc2ce4
Dist. integration tests for select formats
For now, only formats that already have I/O tests available. Most don't
run in CI yet because of missing test data and infra (#86). Includes
test cases for ClusterDataSet edge cases.
