### 2020-01-13 8bcf82e
codeclimate: reduce complexity of DMDataSet._get_fileset
