### 2020-01-13 94257cb
codeclimate: convert some asserts in ClusterDataSet
