### 2020-01-13 e511ed7
simplify ClusterDataSet.check_valid
