+++ docs/source/changelog/bugfix/blobfinder.rst
@@ -1,0 +1,4 @@
+[Bugfix] Bounds checking in blobfinder
+======================================
+
+ * Fix bounds checking, size and index calculation bugs (:issue:`539`, :pr:`548`)
