### 2020-01-14 e707d02
Allow running with NUMBA_JIT_DISABLED=1; Bugfixes
* Substitute to_fixed_tuple() intrinsic with equivalent function if JIT is disabled
* Fix index and size miscalculation bugs in block-wise corelation code
* Add corresponding test cases

Closes #539

@sk1p thx for adding the facilities to run with JIT disabled. It turns out that Numba is
sensitive to the same kind of illegal memory acces bugs as C and only running it with JIT disabled
catches this. :thumbsup:
