### 2020-01-14 0ee9009
Updating changelog items
* Add tags to improve readability of "continuous" changelog
* Promote usability improvements from bugfix to new feature
