+++ docs/source/changelog/features/dismiss_err_via_keys.rst
@@ -1,4 +1,4 @@
-Dismiss error messages via keyboard
-===================================
+[Feature] Dismiss error messages via keyboard
+=============================================
 
  * Allows pressing the escape key to close all currently open error messages (:issue:`437`)
