+++ docs/source/changelog/bugfix/blobfinder.rst
@@ -1,4 +1,5 @@
 [Bugfix] Bounds checking in blobfinder
 ======================================
 
- * Fix bounds checking, size and index calculation bugs (:issue:`539`, :pr:`548`)
+ * Fix bounds checking, size and index calculation bugs in blobfinder (:issue:`539`, :pr:`548`)
+ 
