+++ docs/source/changelog/bugfix/fix_pick_roi_interaction.rst
@@ -1,4 +1,0 @@
-Disable ROI selector if the current mode is PICK
-================================================
-
- * ROI doesn't have any effect if in pick mode, so we hide the dropdown in that case (:issue:`511`)
