### 2020-01-14 f202305
Add tag to HDFS obsolescence changelog entry
