### 2020-01-15 c9fc491
Parametrize and fix limits
* Parametrize instead of for loop, thx @sk1p!
* Fix bug in limit calculation to catch off-by-one type of errors
