### 2020-01-16 5ff04f3
Move MaskContainer from libertem.job.masks to libertem.common.container
The Slice information is available directly in UDFs, which means that
we can keep the dependencies on dataset specifics out of the container.

A sub-class with the previous functionality is implemented for
compatibility with ApplyMasksJob
