+++ src/libertem/job/masks.py
@@ -1,4 +1,3 @@
-import functools
 import logging
 
 try:
@@ -8,51 +7,33 @@ except ImportError:
 import sparse
 import scipy.sparse
 import numpy as np
-import cloudpickle
 
 from libertem.io.dataset.base import DataTile, Partition
 from .base import BaseJob, Task, ResultTile
-from libertem.masks import to_dense, to_sparse, is_sparse
 from libertem.common import Slice
 from libertem.common.buffers import zeros_aligned
+from libertem.common.container import MaskContainer
+
 
 log = logging.getLogger(__name__)
 
 
-def _make_mask_slicer(computed_masks, dtype, sparse_backend, transpose):
-    @functools.lru_cache(maxsize=None)
-    def _get_masks_for_slice(slice_):
-        stack_height = computed_masks.shape[0]
-        m = slice_.get(computed_masks, sig_only=True)
-        # We need the mask's signal dimension flattened
-        m = m.reshape((stack_height, -1))
-        if transpose:
-            # We need the stack transposed in the next step
-            m = m.T
-
-        if is_sparse(m):
-            if sparse_backend == 'sparse.pydata':
-                # sparse.pydata.org is fastest for masks with few layers
-                # and few entries
-                return m.astype(dtype)
-            elif 'scipy.sparse' in sparse_backend:
-                # Just for calculation: scipy.sparse.csr_matrix is
-                # the fastest for dot product of deep mask stack
-                iis, jjs = m.coords
-                values = m.data
-                if sparse_backend == 'scipy.sparse.csc':
-                    return scipy.sparse.csc_matrix((values, (iis, jjs)), shape=m.shape, dtype=dtype)
-                else:
-                    return scipy.sparse.csr_matrix((values, (iis, jjs)), shape=m.shape, dtype=dtype)
-            else:
-                raise ValueError(
-                    "sparse_backend %s not implemented, can be 'scipy.sparse', "
-                    "'scipy.sparse.csc' or 'sparse.pydata'" % sparse_backend)
+class JobMaskContainer(MaskContainer):
+    def get(self, key, dtype=None, sparse_backend=None, transpose=True):
+        if isinstance(key, Partition):
+            slice_ = key.slice
+        elif isinstance(key, DataTile):
+            slice_ = key.tile_slice
+        elif isinstance(key, Slice):
+            slice_ = key
         else:
-            # We convert to the desired type.
-            # This makes sure it is in row major, dense layout as well
-            return m.astype(dtype)
-    return _get_masks_for_slice
+            raise TypeError(
+                "JobMaskContainer.get() can only be called with "
+                "DataTile/Slice/Partition instances"
+            )
+        return super().get(
+            key=slice_, dtype=dtype, sparse_backend=sparse_backend, transpose=transpose
+        )
 
 
 class ApplyMasksJob(BaseJob):
@@ -73,7 +54,7 @@ class ApplyMasksJob(BaseJob):
         # If None, decide in the mask container
         if use_sparse is True:
             use_sparse = 'scipy.sparse'
-        self.masks = MaskContainer(mask_factories, dtype=mask_dtype,
+        self.masks = JobMaskContainer(mask_factories, dtype=mask_dtype,
             use_sparse=use_sparse, count=mask_count)
 
         self.dtype = dtype
@@ -111,165 +92,6 @@ class ApplyMasksJob(BaseJob):
             return self.dtype
 
 
-class MaskContainer(object):
-    '''
-    Container for mask stacks that are created from factory functions.
-
-    It allows stacking, cached slicing, transposing and conversion
-    to condition the masks for high-performance dot products.
-    '''
[…]
