### 2020-01-16 a83e458
Test for UDF dtype logic
