### 2020-01-16 a00d8e8
Towards replacing ApplyMasksJob
* Implement ApplyMasksUDF
* Allow MasksAnalysis to use both Job and UDF back-end interchangeably
* Run all tests in tests_analysis_masks with both back-ends, as far as possible
* Documentation for dtype logic
* Depreciation info for Job API in changelog, to be extended.
