+++ docs/source/changelog/misc/udfmasks.rst
@@ -1,0 +1,5 @@
+[Misc] Optional UDF interface for MasksAnalysis
+===============================================
+
+ * Add necessary infrastructure to use MasksAnalysis with ApplyMasksUDF instead of ApplyMasksJob back-end (:issue:`549`, :pr:`550`)
+ 
