+++ docs/source/reference/udf.rst
@@ -53,3 +53,7 @@ Some generally useful UDFs are included with LiberTEM:
 .. autoclass:: libertem.udf.sumsigudf.SumSigUDF
     :members:
     :special-members: __init__
+
+.. autoclass:: libertem.udf.masks.ApplyMasksUDF
+    :members:
+    :special-members: __init__
