+++ docs/source/changelog/features/dtypes.rst
@@ -1,0 +1,4 @@
+[Feature] dtype support for UDFs
+================================
+
+*  dtype support for UDFs :ref:`udf dtype` (:issue:`549`, :pr:`550`)
