+++ docs/source/changelog/features/udfmasks.rst
@@ -1,0 +1,4 @@
+[Feature] ApplyMasksUDF
+=======================
+
+ * Implement :class:`libertem.udf.masks.ApplyMasksUDF` as a replacement of ApplyMasksJob (:issue:`549`, :pr:`550`)
