+++ tests/test_analysis_masks.py
@@ -2,23 +2,24 @@ import pytest
 import numpy as np
 import scipy.sparse as sp
 import sparse
-from libertem.masks import to_dense, to_sparse, is_sparse
-from libertem.io.dataset.memory import MemoryDataSet
 
 from utils import _naive_mask_apply, _mk_random
 
 import libertem.api as api
+from libertem.masks import to_dense, to_sparse, is_sparse
+from libertem.io.dataset.memory import MemoryDataSet
+from libertem.udf.masks import ApplyMasksUDF
 
 
-def _run_mask_test_program(lt_ctx, dataset, mask, expected):
+def _run_mask_test_program(lt_ctx, dataset, mask, expected, use_udf=False):
     analysis_default = lt_ctx.create_mask_analysis(
-        dataset=dataset, factories=[lambda: mask]
+        dataset=dataset, factories=[lambda: mask], _use_udf=use_udf
     )
     analysis_sparse = lt_ctx.create_mask_analysis(
-        dataset=dataset, factories=[lambda: to_sparse(mask)], use_sparse=True
+        dataset=dataset, factories=[lambda: to_sparse(mask)], use_sparse=True, _use_udf=use_udf
     )
     analysis_dense = lt_ctx.create_mask_analysis(
-        dataset=dataset, factories=[lambda: to_dense(mask)], use_sparse=False
+        dataset=dataset, factories=[lambda: to_dense(mask)], use_sparse=False, _use_udf=use_udf
     )
     results_default = lt_ctx.run(analysis_default)
     results_sparse = lt_ctx.run(analysis_sparse)
@@ -72,68 +73,96 @@ def test_weird_partition_shapes_1_fast(lt_ctx):
     assert tuple(t.tile_slice.shape) == (1, 8, 8, 8)
 
 
-def test_normal_partition_shape(lt_ctx):
+@pytest.mark.parametrize(
+    'use_udf',
+    [True, False]
+)
+def test_normal_partition_shape(lt_ctx, use_udf):
     data = _mk_random(size=(16, 16, 16, 16), dtype="<u2")
     mask = _mk_random(size=(16, 16))
     expected = _naive_mask_apply([mask], data)
 
     dataset = MemoryDataSet(data=data, tileshape=(1, 16, 16), num_partitions=2)
 
-    _run_mask_test_program(lt_ctx, dataset, mask, expected)
+    _run_mask_test_program(lt_ctx, dataset, mask, expected, use_udf)
 
 
-def test_single_frame_tiles(lt_ctx):
+@pytest.mark.parametrize(
+    'use_udf',
+    [True, False]
+)
+def test_single_frame_tiles(lt_ctx, use_udf):
     data = _mk_random(size=(16, 16, 16, 16), dtype="<u2")
     mask = _mk_random(size=(16, 16))
     expected = _naive_mask_apply([mask], data)
 
     dataset = MemoryDataSet(data=data, tileshape=(1, 16, 16), num_partitions=2)
 
-    _run_mask_test_program(lt_ctx, dataset, mask, expected)
+    _run_mask_test_program(lt_ctx, dataset, mask, expected, use_udf)
 
 
+@pytest.mark.parametrize(
+    'use_udf',
+    [True, False]
+)
 @pytest.mark.slow
-def test_subframe_tiles_slow(lt_ctx):
+def test_subframe_tiles_slow(lt_ctx, use_udf):
     data = _mk_random(size=(16, 16, 16, 16), dtype="<u2")
     mask = _mk_random(size=(16, 16))
[…]
