+++ docs/source/changelog/obsolescence/jobremoval.rst
@@ -1,0 +1,6 @@
+[Depreciation] Job API will be retired
+======================================
+
+ * ApplyMasksJob and PickFrameJob will be replaced with implementations based on
+   the UDF interface. The Job API and related functionality will be retired in a
+   future release after a reasonable transition period, details to follow. (:issue:`549`, :pr:`550`).

[…]
