### 2020-01-17 f6eee77
Use UDF interface by default for create_mask_analysis
Tests show that the UDF version is faster for relevant use cases #550
