### 2020-01-17 b02e885
Remove duplication in get_result / get_udf_result
Resolve test bug introduced by using UDFs by default
