### 2020-01-17 7529d49
Move all mask-based Analyses to use UDF by default
