+++ src/libertem/analysis/com.py
@@ -55,6 +55,9 @@ class COMResultSet(AnalysisResultSet):
 
 
 class COMAnalysis(BaseMasksAnalysis):
+    TYPE = 'UDF'
+
+    # FIXME remove this after UDF version is final
     def get_results(self, job_results):
         shape = tuple(self.dataset.shape.nav)
         img_sum, img_x, img_y = (
@@ -62,6 +65,18 @@ class COMAnalysis(BaseMasksAnalysis):
             job_results[1].reshape(shape),
             job_results[2].reshape(shape)
         )
+        return self.get_generic_results(img_sum, img_x, img_y)
+
+    def get_udf_results(self, udf_results, roi):
+        data = udf_results['intensity'].data
+        img_sum, img_x, img_y = (
+            data[..., 0],
+            data[..., 1],
+            data[..., 2],
+        )
+        return self.get_generic_results(img_sum, img_x, img_y)
+
+    def get_generic_results(self, img_sum, img_x, img_y):
         ref_x = self.parameters["cx"]
         ref_y = self.parameters["cy"]
         x_centers = np.divide(img_x, img_sum, where=img_sum != 0)
