+++ src/libertem/analysis/point.py
@@ -7,9 +7,19 @@ from .masks import BaseMasksAnalysis, SingleMaskResultSet
 
 
 class PointMaskAnalysis(BaseMasksAnalysis):
+    TYPE = 'UDF'
+
+    # FIXME remove this after UDF version is final
     def get_results(self, job_results):
         shape = tuple(self.dataset.shape.nav)
         data = job_results[0].reshape(shape)
+        return self.get_generic_results(data)
+
+    def get_udf_results(self, udf_results, roi):
+        data = udf_results['intensity'].data
+        return self.get_generic_results(data[..., 0])
+
+    def get_generic_results(self, data):
         if data.dtype.kind == 'c':
             return SingleMaskResultSet(
                 self.get_complex_results(
