### 2020-01-17 b4c77b4
Move all mask-based Analyses to use UDF by default
