+++ src/libertem/udf/raw.py
@@ -1,0 +1,48 @@
+import logging
+
+import numpy as np
+
+from libertem.udf import UDF
+
+
+log = logging.getLogger(__name__)
+
+
+class PickUDF(UDF):
+    def get_preferred_input_dtype(self):
+        # We load the native dtype
+        return bool
+
+    def get_result_buffers(self):
+        dtype = self.meta.input_dtype
+        sigshape = tuple(self.meta.dataset_shape.sig)
+        if self.meta.roi is not None:
+            navsize = np.count_nonzero(self.meta.roi)
+        else:
+            navsize = np.prod(self.meta.dataset_shape.nav)
+        warn_limit = 2**28
+        loaded_size = np.prod(sigshape) * navsize * np.dtype(dtype).itemsize
+        if loaded_size > warn_limit:
+            log.warning("PickUDF is loading %s bytes, exceeding warning limit %s. \
+                Consider using or implementing an UDF to process data on the worker \
+                nodes instead." % (loaded_size, warn_limit))
+        # We are using a "single" buffer since we mostly load single frames. A
+        # "sig" buffer would work as well, but would require a transpose to
+        # accomodate multiple frames in the last and not first dimension.
+        # A "nav" buffer would allocate a NaN-filled buffer for the whole dataset.
+        return {
+            'intensity': self.buffer(
+                kind='single', extra_shape=(navsize, ) + sigshape, dtype=dtype
+            )
+        }
+
+    def process_tile(self, tile):
+        # We work in flattened nav space with ROI applied
+        sl = self.meta.slice.get()
+        self.results.intensity[sl] = tile
+
+    def merge(self, dest, src):
+        # We receive full-size buffers from each node that
+        # contributes at least one frame and rely on the rest being filled
+        # with zeros correctly.
+        dest['intensity'][:] += src['intensity']
