+++ docs/source/changelog/features/udfmasks.rst
@@ -1,4 +1,6 @@
-[Feature] ApplyMasksUDF
-=======================
+[Feature] ApplyMasksUDF and PickUDF
+===================================
 
  * Implement :class:`libertem.udf.masks.ApplyMasksUDF` as a replacement of ApplyMasksJob (:issue:`549`, :pr:`550`)
+ * Implement :class:`libertem.udf.raw.PickUDF` as a replacement of PickFrameJob (:issue:`549`, :pr:`550`)
+ 
