### 2020-01-17 1a1cda9
Replace PickFrameJob functionality with PickUDF
For now, allow using both by changing the TYPE for comparison purposes.

The UDF version feels slightly more responsive in the GUI.
