+++ tests/test_analysis_raw.py
@@ -77,7 +77,10 @@ def test_get_multiple_frames_squeeze():
     assert np.allclose(result[0:2], data[0, 5:7])
 
 
-def test_pick_analysis(lt_ctx):
+@pytest.mark.parametrize(
+    'TYPE', ['JOB', 'UDF']
+)
+def test_pick_analysis(lt_ctx, TYPE):
     """
     the other tests cover the pick job, this one uses the analysis
     """
@@ -90,13 +93,17 @@ def test_pick_analysis(lt_ctx):
     )
 
     analysis = PickFrameAnalysis(dataset=dataset, parameters={"x": 5, "y": 5})
+    analysis.TYPE = TYPE
     result = lt_ctx.run(analysis)
 
     assert result.intensity.raw_data.shape == (16, 16)
     assert np.allclose(result.intensity.raw_data, data[5, 5])
 
 
-def test_pick_from_3d_ds(lt_ctx):
+@pytest.mark.parametrize(
+    'TYPE', ['JOB', 'UDF']
+)
+def test_pick_from_3d_ds(lt_ctx, TYPE):
     data = _mk_random(size=(16 * 16, 16, 16))
     dataset = MemoryDataSet(
         data=data,
@@ -106,6 +113,7 @@ def test_pick_from_3d_ds(lt_ctx):
     )
 
     analysis = PickFrameAnalysis(dataset=dataset, parameters={"x": 5})
+    analysis.TYPE = TYPE
     result = lt_ctx.run(analysis)
 
     assert result.intensity.raw_data.shape == (16, 16)
@@ -160,7 +168,10 @@ def test_pick_from_3d_ds_job_w_shape_2(lt_ctx):
     assert np.allclose(result, data[8])
 
 
-def test_pick_analysis_via_api_1(lt_ctx):
+@pytest.mark.parametrize(
+    'TYPE', ['JOB', 'UDF']
+)
+def test_pick_analysis_via_api_1(lt_ctx, TYPE):
     data = _mk_random(size=(16, 16, 16, 16))
     dataset = MemoryDataSet(
         data=data,
@@ -170,13 +181,17 @@ def test_pick_analysis_via_api_1(lt_ctx):
     )
 
     analysis = lt_ctx.create_pick_analysis(dataset=dataset, x=8, y=7)
+    analysis.TYPE = TYPE
     result = lt_ctx.run(analysis)
 
     assert result.intensity.raw_data.shape == (16, 16)
     assert np.allclose(result.intensity.raw_data, data[7, 8])
 
 
-def test_pick_analysis_via_api_2_3d_ds(lt_ctx):
+@pytest.mark.parametrize(
+    'TYPE', ['JOB', 'UDF']
+)
+def test_pick_analysis_via_api_2_3d_ds(lt_ctx, TYPE):
     data = _mk_random(size=(16 * 16, 16, 16))
     dataset = MemoryDataSet(
         data=data,
@@ -186,13 +201,17 @@ def test_pick_analysis_via_api_2_3d_ds(lt_ctx):
     )
 
     analysis = lt_ctx.create_pick_analysis(dataset=dataset, x=8)
+    analysis.TYPE = TYPE
     result = lt_ctx.run(analysis)
 
     assert result.intensity.raw_data.shape == (16, 16)
     assert np.allclose(result.intensity.raw_data, data[8])
 
 
-def test_pick_analysis_via_api_3_3d_ds_fail_1(lt_ctx):
+@pytest.mark.parametrize(
+    'TYPE', ['JOB', 'UDF']
+)
+def test_pick_analysis_via_api_3_3d_ds_fail_1(lt_ctx, TYPE):
     data = _mk_random(size=(16 * 16, 16, 16))
[…]
