### 2020-01-17 be3fd48
Fix inheritance and switching logic
