### 2020-01-17 a52d42e
Fix inheritance and switching logic
