### 2020-01-17 0b6cbab
Enable memory mapping for UDFs
Is there a reason against using this?
