### 2020-01-18 d706dfc
Polish for documentation
* Fix misnamed parameter
* Hide superclass docstrings
* Include PickUDF in UDF reference docs
