### 2020-01-18 1e665db
Polish for documentation
* Fix misnamed parameter
* Hide superclass docstrings
* Include PickUDF in UDF reference docs
