### 2020-01-20 LiberTEM/LiberTEM#554 @uellue
Tile size and read mode negotiation for UDFs
In relation to #549 and #551: The optimal tile size for `process_tile()` depends on the circumstances:

* The UDF creates intermediate data similar in size to the tile (log scaling etc), the dataset doesn't use mmap or the data is converted before passing to the UDF: Tiles of approx. 1 MB are optimal.
* The dataset can return mmapped data directly without any conversion and the UDF works efficiently on large tiles, for example with a pure dot product with a shallow mask stack: Very large tiles up to partition size can be optimal.

For that reason, the UDF interface should allow for negotiation of the tile size.

Idea by @sk1p: Instead of directly using `get_tiles()` with loads of parameters, the partition could allow to create a ProtoTileStream object which handles the decision logic on Dataset side (full frame? mmap? dtype?) and can give information on the actual read strategy with the given parameters before tiles are loaded.

The UDF can indicate a preferred upper limit for tiles, which could be 1 MB by default. This is in no way guaranteed since reading full K2 frames will be way above this limit.

The UDFRunner can then decide on a preferred tile size based on the ProtoTileStream, UDF method and UDF preferred size information.
