### 2020-01-20 e7c2348
Introduce a constant for using native dataset dtype
Improve readability, thx @sk1p
