### 2020-01-20 5e134a0
Introduce a constant for using native dataset dtype
Improve readability, thx @sk1p
