### 2020-01-20 2b1a5ea
Remove code duplication
