### 2020-01-20 3889231
Remove code duplication
