### 2020-01-20 0c9332f
Remove _use_udf parameter
Set TYPE instead for debugging and testing
