+++ tests/test_analysis_masks.py
@@ -11,16 +11,19 @@ from libertem.io.dataset.memory import MemoryDataSet
 from libertem.udf.masks import ApplyMasksUDF
 
 
-def _run_mask_test_program(lt_ctx, dataset, mask, expected, use_udf=False):
+def _run_mask_test_program(lt_ctx, dataset, mask, expected, TYPE='JOB'):
     analysis_default = lt_ctx.create_mask_analysis(
-        dataset=dataset, factories=[lambda: mask], _use_udf=use_udf
+        dataset=dataset, factories=[lambda: mask]
     )
+    analysis_default.TYPE = TYPE
     analysis_sparse = lt_ctx.create_mask_analysis(
-        dataset=dataset, factories=[lambda: to_sparse(mask)], use_sparse=True, _use_udf=use_udf
+        dataset=dataset, factories=[lambda: to_sparse(mask)], use_sparse=True
     )
+    analysis_sparse.TYPE = TYPE
     analysis_dense = lt_ctx.create_mask_analysis(
-        dataset=dataset, factories=[lambda: to_dense(mask)], use_sparse=False, _use_udf=use_udf
+        dataset=dataset, factories=[lambda: to_dense(mask)], use_sparse=False
     )
+    analysis_dense.TYPE = TYPE
     results_default = lt_ctx.run(analysis_default)
     results_sparse = lt_ctx.run(analysis_sparse)
     results_dense = lt_ctx.run(analysis_dense)
@@ -74,95 +77,95 @@ def test_weird_partition_shapes_1_fast(lt_ctx):
 
 
 @pytest.mark.parametrize(
-    'use_udf',
-    [True, False]
+    'TYPE',
+    ['JOB', 'UDF']
 )
-def test_normal_partition_shape(lt_ctx, use_udf):
+def test_normal_partition_shape(lt_ctx, TYPE):
     data = _mk_random(size=(16, 16, 16, 16), dtype="<u2")
     mask = _mk_random(size=(16, 16))
     expected = _naive_mask_apply([mask], data)
 
     dataset = MemoryDataSet(data=data, tileshape=(1, 16, 16), num_partitions=2)
 
-    _run_mask_test_program(lt_ctx, dataset, mask, expected, use_udf)
+    _run_mask_test_program(lt_ctx, dataset, mask, expected, TYPE)
 
 
 @pytest.mark.parametrize(
-    'use_udf',
-    [True, False]
+    'TYPE',
+    ['JOB', 'UDF']
 )
-def test_single_frame_tiles(lt_ctx, use_udf):
+def test_single_frame_tiles(lt_ctx, TYPE):
     data = _mk_random(size=(16, 16, 16, 16), dtype="<u2")
     mask = _mk_random(size=(16, 16))
     expected = _naive_mask_apply([mask], data)
 
     dataset = MemoryDataSet(data=data, tileshape=(1, 16, 16), num_partitions=2)
 
-    _run_mask_test_program(lt_ctx, dataset, mask, expected, use_udf)
+    _run_mask_test_program(lt_ctx, dataset, mask, expected, TYPE)
 
 
 @pytest.mark.parametrize(
-    'use_udf',
-    [True, False]
+    'TYPE',
+    ['JOB', 'UDF']
 )
 @pytest.mark.slow
-def test_subframe_tiles_slow(lt_ctx, use_udf):
+def test_subframe_tiles_slow(lt_ctx, TYPE):
     data = _mk_random(size=(16, 16, 16, 16), dtype="<u2")
     mask = _mk_random(size=(16, 16))
     expected = _naive_mask_apply([mask], data)
 
[…]
