### 2020-01-20 861049a
Refactor to reduce cognitive complexity
