### 2020-01-20 9fe56ec
General function to calculate origin and check for parameters
