### 2020-01-20 9faa45c
General function to calculate origin and check for parameters
