### 2020-01-20 LiberTEM/LiberTEM#554 @sk1p
I really like this, as it allows us to finally remove the bespoke `tileshape` params and some of the  format-specific `tileshape` calculations... it also means we have a central point where we can tweak the tilesize, possibly allowing to make it larger for machines with huge L3 cache (latest AMD Ryzens have up to 256MB L3, that is about 4MB L3 cache **per core**)

So, instead of `get_tiles` on the `Partition`, we would have `get_tilestream` with the same parameters:

```python
def get_tilestream(self, crop_to=None, full_frames=False, mmap=False, dest_dtype="float32",
                   roi=None, target_size=None):
    # ...
    return TileStream(...)
```

The `TileStream` can then be queried for the concrete reading modes (estimated tile size, mmap enabled?, ...) and parameters can be updated accordingly.

```python
# example usage:
stream = partition.get_tilestream(full_frames=True, roi=roi, dest_dtype=dest_dtype)
stream.set_tilesize(min(stream.get_max_tilesize(), udf.get_the_preferred_max_tile_size()))

# then, either of these forms could be implemented:
tiles = stream.get_tiles()
for tile in tiles:
    pass

for tile in tilestream:
    pass
# in the second form, setting a parameter after starting iteration should raise an exception,
# which is implicitly handled in the first form (only after calling `get_tiles` again will the new parameters take effect)
# the name fits better for the second form, IMHO

# also possible: builder-like parameter passing, which is natural for methods with
# many parameters and default values for most of them:
stream = partition.get_tilestream()
stream.set_roi(roi)
stream.set_dest_dtype(dest_dtype)
if method == 'frame':
    stream.set_fullframe(True)
for tile in stream:
    ...
```

Nice side-effect: a `TileStream` can then also contain information about the concrete tiling scheme that will be used for the given parameters - which can be interesting for constructing buffer structures that mirror the given tiling scheme.

Remaining questions:

 * Should there be symbolic values for "whatever fits in L3" or "as large as possible", or other values?
 * What does the UDF need as input to decide its maximum tile size? The mask UDF would be happy with returning "as large as possible" without any input, right? In case of mmap, this would be granted, otherwise, the data set would limit it to L3-sized tiles
 * Does it need to know anything about reading modes (mmap, direct, ...)? I guess this could be some kind of namespace (like `UDFMeta`), which can be filled with items as needed
