### 2020-01-21 7e97027
Increase number of function arguments
Refs #541
