### 2020-01-21 bc3871b
Reference new location of MaskContainer; don't reference ApplyMaskJob anymore
Thx @sk1p!
