### 2020-01-21 c9f67d1
assorted doc fixes
