### 2020-01-21 LiberTEM/LiberTEM#554 @uellue
Yes, symbolic values would be great, similar to `UDF.USE_NATIVE_DTYPE` in order to eliminate opaque magic numbers. Plus, we might be able to tune them for the system at a later time -- AMD EPYC and such!

Interesting aspects from #247 and #310: Not only the size, but also the shape of a tile is relevant. An interface to indicate the preferred number of frames in a tile could be nice. The ideal for numba-based processing is somewhere around 8 frames to use the L1 and L2 cache efficiently in operations like #398.

Probably a dataset could decide then in what way it would split the individual frames since that doesn't matter that much anymore as long as the tiles contain sequential pieces of memory that are "long enough" (a block size of 128 sequential numbers is more than sufficient in my experience). The dataset should know best what way of splitting frames makes sense and produces nice stretches of sequential access.

Some datasets could be very flexible in their tiling scheme, namely RAW. Others like K2IS can only offer a limited set of tilings. Plus, we might have situations where a dataset can offer mmap only for specific tiling schemes -- imagine K2IS with uint16 or float32 instead of packed uint12!

Perhaps the UDF should then pick the tiling scheme that it finds most suitable?

The process could be then as following:

* UDF indicates preferred tile size (soft), preferred frame count (soft), preferred dtype (hard lower bound) and ability to process tiled data through implementing `process_tile()` or not (hard).
* ProtoTileStream generates offers. If it is very restricted like K2IS, it will return items from a short  list of fixed options, filtered by the hard constraints, with "easy" ones first. If it is flexible like RAW, it will generate one best matching option for the constraints . The UDF then picks one, with default operation of picking the first.
