### 2020-01-21 LiberTEM/LiberTEM#554 @uellue
> * Does it need to know anything about reading modes (mmap, direct, ...)? I guess this could be some kind of namespace (like `UDFMeta`), which can be filled with items as needed

Perhaps the information could be available. The main point here is to decide on an upper bound for tile size if the data has to be converted first, right? It could be sufficient to let the ProtoTileStream apply its own upper constraint on tile size depending on the reading method.
