### 2020-01-21 LiberTEM/LiberTEM#554 @uellue
> I really like this, as it allows us to finally remove the bespoke `tileshape` params and some of the format-specific `tileshape` calculations...

Yes! :-) :-) :-)
