### 2020-01-21 LiberTEM/LiberTEM#554 @uellue
Ah, and we are also getting close to adaptive partition sizes #382!
