### 2020-01-21 LiberTEM/LiberTEM#554 @sk1p
> Some datasets could be very flexible in their tiling scheme, namely RAW.

Should clarify this - when tiling in the signal dimension, even the RAW reader has quite some limits, imposed by the actual I/O method. For example, reading "thin" columns means we a) have high overhead in the number of `read` syscalls (when not using liburing), and b) read amplification because block sizes (`PAGE_SIZE` etc.). The read amplification issue is not so bad I think (if those blocks aren't touched by the CPU but stay in page cache), but large number of syscalls could kill performance quite reliably.
