### 2020-01-21 bbfaf2c
Start updating for GSoC 2020
Refs #543, see discussion!
