+++ docs/source/gsoc.rst
@@ -1,46 +1,90 @@
-GSoC 2019 ideas
+GSoC 2020 ideas
 ===============
 
-LiberTEM is participating in the `Google Summer of Code <https://summerofcode.withgoogle.com/>`_ as a sub-organization of the `Python Software Foundation <https://python-gsoc.org/>`_. As a student, you can get paid by Google for three months, have fun working on an interesting open source software project, gain real-world development experience, and do something that looks nice on your CV!
+.. note::
+    Our application for GSoC 2020 is still work in progress, but you can already
+    contact us about potential projects! Just be aware that our participation is
+    not officially confirmed, applications are not open yet and this page is
+    currently under active development.
+
+LiberTEM is participating in the `Google Summer of Code
+<https://summerofcode.withgoogle.com/>`_ as a sub-organization of the `Python
+Software Foundation <https://python-gsoc.org/>`_. As a student, you can get paid
+by Google for three months, have fun working on an interesting open source
+software project, gain real-world development experience, and do something that
+looks nice on your CV!
 
 * Check out our description and project ideas below
 * Contact us if you'd like to work on LiberTEM
 * Prepare a `proposal <https://python-gsoc.org/index.html#apply>`_ together with us
-* You submit your application at the Google Summer of Code homepage to the Python Software Foundation organization, naming LiberTEM as the sub-organization.
+* You submit your application at the Google Summer of Code homepage to the Python
+  Software Foundation organization, naming LiberTEM as the sub-organization.
 
 Why LiberTEM
 --------------
 
-`LiberTEM <.>`_ is an open source platform for high-throughput distributed processing of pixelated scanning transmission electron microscopy (STEM) data. It is created to deal with the terabytes of data that modern high-speed high-resolution detectors for electron microscopy can produce. Our :doc:`architecture` page describes in more detail how exactly it works.
+`LiberTEM <.>`_ is an open source platform for high-throughput distributed
+processing of pixelated scanning transmission electron microscopy (STEM) data.
+It is created to deal with the terabytes of data that modern high-speed
+high-resolution detectors for electron microscopy can produce. Our
+:doc:`architecture` page describes in more detail how exactly it works.
 
 ..  figure:: ./images/Principle.png
     :scale: 50%
     :alt: In pixelated STEM, a full diffraction image is recorded for each scan position.
 
-    *In pixelated STEM, a sample is scanned with a focused electron beam, and a full image of the transmitted beam is recorded for each scan position. The result is a four-dimensional data hypercube. This application can generate tremendous amounts of data from high-resolution scans with a high-speed high-resolution detector.*
+    *In pixelated STEM, a sample is scanned with a focused electron beam, and a
+    full image of the transmitted beam is recorded for each scan position. The
+    result is a four-dimensional data hypercube. This application can generate
+    tremendous amounts of data from high-resolution scans with a high-speed
+    high-resolution detector.*
 
-The project started in the beginning of 2018 and is currently attracting more and more users because it is orders of magnitude faster and more scalable than established solutions, and at the same time convenient to use. In our field, LiberTEM is seen as the up and coming technology to tap the full scientific potential of modern high-speed detectors. Such detectors can already reach a data rate of 8.5 GB/s and will soon produce more than 50 GB/s of raw data. The conventional established PC-based solutions cannot keep up with such data rates, and distributed systems like LiberTEM are required for progress in this field.
+The project started in the beginning of 2018 and is currently attracting more
+and more users because it is orders of magnitude faster and more scalable than
+established solutions, and at the same time convenient to use. In our field,
+LiberTEM is seen as the up and coming technology to tap the full scientific
[…]
