### 2020-01-21 59a9012
Add profiling instructions to docs, fixes #552
