+++ docs/source/udf/profiling.rst
@@ -1,0 +1,180 @@
+.. _`udf profiling`:
+
+.. testsetup:: *
+
+    from libertem.api import Context
+
+
+Profiling UDFs
+==============
+
+To bring your UDF to production, it may be necessary to look into the performance and
+efficiency of your function. There may be some low-hanging fruit to optimize away, but
+they only really become visible once you profile your code. Just looking at code and guessing
+which parts are expensive can be misleading and cause you to spend time optimizing the wrong part
+of your code. Always measure!
+
+Profiling means instrumenting the execution of a program and finding out which parts
+of your code use how much of a given resource (for example, CPU time, or memory). 
+
+Prerequisite: inline executor
+-----------------------------
+
+By default, your UDF will be run in a multi-process or multi-threaded environment. This makes
+profiling a challenge. But as most of the time, improving the execution time of your UDF in a
+single-threaded executor also improves the respective multi-process execution time, we can run
+the UDF in a single thread to do profiling measurements.
+
+LiberTEM provides what it calls :class:`~libertem.executor.inline.InlineJobExecutor`,
+which executes your UDF in a single thread, thus allowing the use of default Python profiling
+mechanisms. 
+
+To use it, pass it to the :class:`~libertem.api.Context` on construction:
+
+.. testcode::
+   
+   from libertem.executor.inline import InlineJobExecutor
+   ctx = Context(executor=InlineJobExecutor())
+
+Then, you can continue as usual, loading data, executing your UDF, etc.
+
+Line profiling using `line_profiler`
+------------------------------------------
+
+Profiling comes in different forms, some will show you how much time each function
+of a whole program took, others can give you a line-by-line overview for functions
+you are interested in, like `line_profiler`. For most UDFs this will be sufficient
+for performance analysis.
+
+First, you need to install the `line_profiler` package via pip, having your
+conda environment or virtualenv activated:
+
+.. code-block:: shell
+
+   (libertem) $ pip install line_profiler
+
+Then the easiest way to get started is to use the ipython/jupyter integration of
+`line_profiler`. Put :code:`%load_ext line_profiler` somewhere in your notebook,
+then you can use the :code:`%lprun` magic command to run your UDF while profiling:
+
+.. code-block::
+
+   %lprun -f YourUDF.get_task_data -f YourUDF.process_frame ctx.run_udf(dataset=dataset, YourUDF())
+
+Note the repeated :code:`-f` arguments - you can list any methods of your UDF you are
+interested in, or even other Python functions you are directly or indirectly using. If you
+had some complicated code in :code:`YourUDF.merge`, you would include :code:`-f YourUDF.merge`
+in the :code:`%lprun` call.
+
+This is how the output could look like:
+
+.. code-block:: text
+
+   Timer unit: 1e-06 s
+
+   Total time: 0.002283 s
+   File: /home/clausen/source/libertem/src/libertem/udf/holography.py
+   Function: get_task_data at line 145
+
+   Line #      Hits         Time  Per Hit   % Time  Line Contents
+   ==============================================================
+      145                                               def get_task_data(self):
+      146                                                   """
+      147                                                   Updates `task_data`
+      148                                           
+      149                                                   Returns
+      150                                                   -------
+      151                                                   kwargs : dict
+      152                                                   A dictionary with the following keys:
+      153                                                       kwargs['aperture'] : array-like
+      154                                                       Side band filter aperture (mask)
+      155                                                       kwargs['slice'] : slice
+      156                                                       Slice for slicing FFT of the hologram
+      157                                                   """
+      158                                           
+      159         2         48.0     24.0      2.1          out_shape = self.params.out_shape
[…]
