### 2020-01-22 #2006 @alvarolopez
@majensen I am done, the paper is ok to go in its current form.
