### 2020-01-22 4af83d6
Profiling docs: review from @uellue
