### 2020-01-22 946ad68
Use the same dtype as _naive_mask_apply
