### 2020-01-22 9b9515c
WebAssembly project
Refs #199, #543
