### 2020-01-23 fce7ee3
Code snippet for GSoC #31
Suggest issue #81 for first PR
