### 2020-01-23 14d58c8
Fix mangled data
Closes #562
