+++ tests/test_analysis_radialfourier.py
@@ -1,6 +1,9 @@
 import pytest
 
+import numpy as np
+
 from libertem.io.dataset.memory import MemoryDataSet
+from libertem.utils.generate import cbed_frame
 
 from utils import _mk_random
 
@@ -16,47 +19,187 @@ def ds_random():
     return dataset
 
 
-def test_smoke(ds_random, lt_ctx):
+@pytest.mark.parametrize(
+    'TYPE', ['JOB', 'UDF']
+)
+def test_smoke(ds_random, lt_ctx, TYPE):
     analysis = lt_ctx.create_radial_fourier_analysis(
         dataset=ds_random, cx=0, cy=0, ri=0, ro=10, n_bins=2, max_order=23
     )
+    analysis.TYPE = TYPE
     results = lt_ctx.run(analysis)
 
 
-def test_smoke_small(ds_random, lt_ctx):
+@pytest.mark.parametrize(
+    'TYPE', ['JOB', 'UDF']
+)
+def test_smoke_small(ds_random, lt_ctx, TYPE):
     analysis = lt_ctx.create_radial_fourier_analysis(
         dataset=ds_random, cx=0, cy=0, ri=0, ro=1, n_bins=1, max_order=23
     )
+    analysis.TYPE = TYPE
     results = lt_ctx.run(analysis)
     # access result to actually create result list:
     results['absolute_0_0']
 
 
-def test_smoke_large(ds_random, lt_ctx):
+@pytest.mark.parametrize(
+    'TYPE', ['JOB', 'UDF']
+)
+def test_smoke_large(ds_random, lt_ctx, TYPE):
     analysis = lt_ctx.create_radial_fourier_analysis(
         dataset=ds_random, cx=0, cy=0, n_bins=1, max_order=23
     )
+    analysis.TYPE = TYPE
     results = lt_ctx.run(analysis)
 
 
-def test_smoke_two(ds_random, lt_ctx):
+@pytest.mark.parametrize(
+    'TYPE', ['JOB', 'UDF']
+)
+def test_smoke_two(ds_random, lt_ctx, TYPE):
     analysis = lt_ctx.create_radial_fourier_analysis(
         dataset=ds_random, cx=0, cy=0, ri=0, ro=2, n_bins=2, max_order=2
     )
+    analysis.TYPE = TYPE
     results = lt_ctx.run(analysis)
     # access result to actually create result list:
     results.complex_1_2
 
 
-def test_smoke_defaults(ds_random, lt_ctx):
+@pytest.mark.parametrize(
+    'TYPE', ['JOB', 'UDF']
+)
+def test_smoke_defaults(ds_random, lt_ctx, TYPE):
     analysis = lt_ctx.create_radial_fourier_analysis(
         dataset=ds_random
     )
+    analysis.TYPE = TYPE
     results = lt_ctx.run(analysis)
 
 
-def test_sparse(ds_random, lt_ctx):
+@pytest.mark.parametrize(
+    'TYPE', ['JOB', 'UDF']
+)
+def test_sparse(ds_random, lt_ctx, TYPE):
     analysis = lt_ctx.create_radial_fourier_analysis(
         dataset=ds_random, use_sparse=True,
     )
+    analysis.TYPE = TYPE
     results = lt_ctx.run(analysis)
+
+
+@pytest.mark.parametrize(
+    'TYPE', ['JOB', 'UDF']
+)
+def test_symmetries(lt_ctx, TYPE):
[…]
