### 2020-01-23 6a63c9f
Fix mangled data
Closes #562
