### 2020-01-23 #2006 @mamcdona77
Will update early next week

Michael
Michael A McDonald, MD, PhD

> On Jan 22, 2020, at 9:52 PM, Mark Jensen <notifications@github.com> wrote:
>
> @alvarolopez - very much appreciated!
> @mamcdona77 - how are things progressing?
>
