+++ docs/source/gsoc.rst
@@ -164,7 +164,7 @@ corresponding data to run the examples!
 Furthermore, you could help us hunt down all references to the Job API in our
 code, introduce depreciation warnings and update documentation where necessary.
 
-Other small and relatively straightforward issues include :issue:`81`
+Other small and relatively straightforward issues include :issue:`81,267,506`
 
 Project ideas
 -------------
@@ -174,7 +174,7 @@ entire time. Please feel free to suggest your own ideas as well! Just working on
 a number of smaller features and getting a good cross-sectional experience of
 LiberTEM can work as well.
 
-0. Beginner: Implement rotation in GUI and back-end for center of mass analysis :issue:`31`
+1. Beginner: Implement rotation in GUI and back-end for center of mass analysis :issue:`31`
     Currently, the implementation assumes that the detector edges are parallel
     to the sample X and Y coordinate directions. This is mostly, but not always
     the case. In this project you implement an additional parameter for rotation
@@ -189,12 +189,13 @@ LiberTEM can work as well.
 
     *Primary contact:* Dieter (@uellue)
 
-0. Beginner: Code snippet for analysis in GUI
+2. Beginner: Code snippet for analysis in GUI
     Currently, our GUI only offers limited capabilities. Most notably, color
     scale, scale bars and exporting results are not implemented. In this
-    project, you implement functions that generate a code snippet ready for copy
-    & paste to use an analysis with the same settings as in the GUI in a script,
-    for example a Jupyter notebook. This allows users more flexibility.
+    project, you implement functions that generate a code snippet in the GUI
+    ready for copy & paste. Pasting this snippet in, for example, a Jupyter
+    notebook allows to use the analysis with the same parameters as in the GUI
+    in a scripting environment that gives users more flexibility.
 
     *Skills:* Python
 
@@ -202,7 +203,7 @@ LiberTEM can work as well.
 
     *Primary contact:* Dieter (@uellue), Alex (@sk1p)
 
-1. Intermediate: Implement an analysis workflow for `RDF mapping <https://publikationen.bibliothek.kit.edu/1000056485/5249497>`_.
+3. Intermediate: Implement an analysis workflow for `RDF mapping <https://publikationen.bibliothek.kit.edu/1000056485/5249497>`_.
     This can give you experience with the product development, design and
     application side of software engineering, and applied data science. A major
     part of the work is first figuring out *what* to implement together with our
@@ -217,7 +218,35 @@ LiberTEM can work as well.
 
     *Primary contact:* Dieter (@uellue)
 
-2. Beginner/Intermediate/Advanced: Compression survey.
+4. Intermediate: Allow reshaping datasets into a custom shape :issue:`441`
+    Data in files is not always stored in the native shape, or different
+    representations may be possible depending on the application. The dataset
+    implementation and GUI should allow specifying a different shape than the
+    layout in the dataset.
+
+    *Skills:* Python, NumPy and TypeScript programming.
+
+    *Domain knowledge:* None
+
+    *Primary contact:* Alex (@sk1p)
+
+
+5. Beginner/Intermediate: Set number of threads and workers dynamically for UDFs :issue:`546`
+    Your task is to implement, document and validate methods to set the thread
+    count for all relevant numerics back-ends in LiberTEM dynamically on a
+    case-by-case basis on the worker processes and main node. Furthermore, you
+    can implement an interface for UDFs to specify their preferred number of
+    threads. This can improve L3 cache efficiency for UDFs that can use a
+    multi-threaded back-end and at the same time avoids oversubscription of the
+    CPU.
+
[…]
