### 2020-01-24 f8cd74b
Test for both JOB and UDF
Make sure that the functionality is tested with the legacy JOB back-end
and the new UDF back-end for the time being
