+++ tests/io/test_frms6.py
@@ -63,8 +63,12 @@ def test_pick_job(default_frms6, lt_ctx):
     assert results.shape == (264, 264)
 
 
-def test_pick_analysis(default_frms6, lt_ctx):
+@pytest.mark.parametrize(
+    'TYPE', ['JOB', 'UDF']
+)
+def test_pick_analysis(default_frms6, lt_ctx, TYPE):
     analysis = PickFrameAnalysis(dataset=default_frms6, parameters={"x": 16, "y": 16})
+    analysis.TYPE = TYPE
     results = lt_ctx.run(analysis)
     assert results[0].raw_data.shape == (264, 264)
 
