+++ tests/io/test_hdf5.py
@@ -14,7 +14,10 @@ from libertem.analysis.sum import SumAnalysis
 from utils import _naive_mask_apply, _mk_random
 
 
-def test_hdf5_apply_masks_1(lt_ctx, hdf5_ds_1):
+@pytest.mark.parametrize(
+    'TYPE', ['JOB', 'UDF']
+)
+def test_hdf5_apply_masks_1(lt_ctx, hdf5_ds_1, TYPE):
     mask = _mk_random(size=(16, 16))
     with hdf5_ds_1.get_reader().get_h5ds() as h5ds:
         data = h5ds[:]
@@ -22,6 +25,7 @@ def test_hdf5_apply_masks_1(lt_ctx, hdf5_ds_1):
     analysis = lt_ctx.create_mask_analysis(
         dataset=hdf5_ds_1, factories=[lambda: mask]
     )
+    analysis.TYPE = TYPE
     results = lt_ctx.run(analysis)
 
     assert np.allclose(
@@ -143,7 +147,10 @@ def test_roi_1(hdf5, lt_ctx):
     assert tiles[0].tile_slice.origin == (0, 0, 0)
 
 
-def test_pick(hdf5, lt_ctx):
+@pytest.mark.parametrize(
+    'TYPE', ['JOB', 'UDF']
+)
+def test_pick(hdf5, lt_ctx, TYPE):
     ds = H5DataSet(
         path=hdf5.filename, ds_path="data", tileshape=(1, 3, 16, 16)
     )
@@ -151,6 +158,7 @@ def test_pick(hdf5, lt_ctx):
     assert len(ds.shape) == 4
     print(ds.shape)
     pick = lt_ctx.create_pick_analysis(dataset=ds, x=2, y=3)
+    pick.TYPE = TYPE
     lt_ctx.run(pick)
 
 
