### 2020-01-27 49e4e10
Set up CI with Azure Pipelines
* Adjust release script, docs etc.
* Publish libertem-dev docker image
* Publish test results and coverage to Azure DevOps
* Adjust event loop policy on Windows

  Apparently Dask.distributed only sets the event loop policy when
  running standalone and not within pytest. For that reason we override
  it to make sure it is always set correctly.

[skip ci]

Co-authored-by: Dieter Weber <d.weber@fz-juelich.de>
