+++ tox.ini
@@ -1,9 +1,9 @@
 [tox]
-envlist = flake8, py{36,37}, benchmark, benchmark-cuda{101,102}, mypy
+envlist = flake8, py{36,37,38,39}, benchmark, benchmark-cuda{101,102}, mypy
 
 [testenv]
 commands=
-    pytest --durations=10 --cov=libertem --cov-report=term --cov-report=html --cov-report=xml {posargs:tests/}
+    pytest --durations=10 --cov=libertem --cov-report=term --cov-report=html --cov-report=xml --junitxml=junit.xml {posargs:tests/}
     # win_tweaks.py depends on modules that are only available on Windows
     pytest --doctest-modules --ignore=src/libertem/win_tweaks.py src/libertem/
 deps=
@@ -27,7 +27,7 @@ passenv=
 
 [testenv:numba_coverage]
 commands=
-    pytest --durations=5 --cov=libertem --cov-report=term --cov-report=html --cov-report=xml -m with_numba {posargs:tests/}
+    pytest --durations=5 --cov=libertem --cov-report=term --cov-report=html --cov-report=xml -m with_numba --junitxml=junit.xml {posargs:tests/}
 setenv=
     NUMBA_DISABLE_JIT=1
 
@@ -106,20 +106,20 @@ deps=
     -roverride_requirements.txt
 skipsdist=True
 
-[testenv:docs-deploy]
+# just a non-watching version of the docs environment above, also building twice for bibs
+[testenv:docs-build-ci]
 changedir={toxinidir}
-basepython=python3.7
+basepython=python3
 setenv=
     PYTHONPATH={toxinidir}
 commands=
     python "{toxinidir}/scripts/build-authors-contributors"
-    # First run to build sphinxcontrib-bibtex index may throw warnings
-    travis-sphinx build --nowarn
-    travis-sphinx build
-    travis-sphinx deploy
+    # Two runs to get complete bibliography. The first one will throw warnings about missing
+    # citations.
+    sphinx-build -b html "{toxinidir}/docs/source" "{toxinidir}/docs/build/html"
+    sphinx-build -W -b html "{toxinidir}/docs/source" "{toxinidir}/docs/build/html"
 deps=
     sphinx
-    travis-sphinx
     sphinxcontrib-bibtex
     sphinx-issues
     nbsphinx
@@ -130,8 +130,6 @@ deps=
 skipsdist=True
 passenv=
     HOME
-    GH_*
-    TRAVIS_*
 
 [testenv:docs-check]
 changedir={toxinidir}
