### 2020-01-29 bd55311
Clean-up, shortening
Feedback from @warthog9, thx!

See also https://github.com/python-gsoc/python-gsoc.github.io/pull/110/
