+++ docs/source/gsoc.rst
@@ -1,12 +1,6 @@
 GSoC 2020 ideas
 ===============
 
-.. note::
-    Our application for GSoC 2020 is still work in progress, but you can already
-    contact us about potential projects! Just be aware that our participation is
-    not officially confirmed, applications are not open yet and this page is
-    currently under active development.
-
 LiberTEM is participating in the `Google Summer of Code
 <https://summerofcode.withgoogle.com/>`_ as a sub-organization of the `Python
 Software Foundation <https://python-gsoc.org/>`_. As a student, you can get paid
@@ -41,12 +35,7 @@ high-resolution detectors for electron microscopy can produce. Our
 
 The project started in the beginning of 2018 and is currently attracting more
 and more users because it is orders of magnitude faster and more scalable than
-established solutions, and at the same time convenient to use. In our field,
-LiberTEM is seen as the up and coming technology to tap the full scientific
-potential of modern high-speed detectors. Such detectors can already reach a
-data rate of 8.5 GB/s and will soon produce more than 50 GB/s of raw data. The
-conventional established PC-based solutions cannot keep up with such data rates,
-and distributed systems like LiberTEM are required for progress in this field.
+established solutions.
 
 Working on LiberTEM will give you experience in developing distributed systems
 for high-performance data processing with Python. You can learn how to profile
@@ -107,10 +96,7 @@ Getting started
 ---------------
 
 If you have questions, please ask freely: Supporting users and contributors has
-a high priority for us and your questions help us improve our documentation. Our
-development is currently moving very quickly. We are planning to complete our
-documentation when a few major features have left the prototype stage. For that
-reason it is always a good idea to be in touch directly.
+a high priority for us and your questions help us improve our documentation.
 
 Installation
 ~~~~~~~~~~~~
@@ -174,7 +160,7 @@ entire time. Please feel free to suggest your own ideas as well! Just working on
 a number of smaller features and getting a good cross-sectional experience of
 LiberTEM can work as well.
 
-1. Beginner: Implement rotation in GUI and back-end for center of mass analysis :issue:`31`
+1. **Beginner**: Implement rotation in GUI and back-end for center of mass analysis :issue:`31`.
     Currently, the implementation assumes that the detector edges are parallel
     to the sample X and Y coordinate directions. This is mostly, but not always
     the case. In this project you implement an additional parameter for rotation
@@ -189,7 +175,7 @@ LiberTEM can work as well.
 
     *Primary contact:* Dieter (@uellue)
 
-2. Beginner: Code snippet for analysis in GUI
+2. **Beginner**: Code snippet for analysis in GUI :issue:`158`.
     Currently, our GUI only offers limited capabilities. Most notably, color
     scale, scale bars and exporting results are not implemented. In this
     project, you implement functions that generate a code snippet in the GUI
@@ -203,7 +189,7 @@ LiberTEM can work as well.
 
     *Primary contact:* Dieter (@uellue), Alex (@sk1p)
 
-3. Intermediate: Implement an analysis workflow for `RDF mapping <https://publikationen.bibliothek.kit.edu/1000056485/5249497>`_.
+3. **Intermediate**: Implement an analysis workflow for `RDF mapping <https://publikationen.bibliothek.kit.edu/1000056485/5249497>`_.
     This can give you experience with the product development, design and
     application side of software engineering, and applied data science. A major
     part of the work is first figuring out *what* to implement together with our
@@ -218,7 +204,7 @@ LiberTEM can work as well.
 
[…]
