### 2020-01-29 060f6cd
Test case for 3D HDF5 file
The HDF5 dataset seems to still have an assumption of being 4D hardcoded.
As an example, libertem.io.dataset.hdf5.H5DataSet.get_partitions() works
with a 4D origin.

The same holds for libertem.io.dataset.base.DataSet.partition_shape() and
libertem.io.utils.get_partition_shape().

LiberTEM should support files with any shape and at least two dimensions.

See also #441

Thx @robertbuecker for reporting!
