+++ conftest.py
@@ -40,6 +40,19 @@ def hdf5(tmpdir_factory):
 
 
 @pytest.fixture(scope='session')
+def hdf5_3D(tmpdir_factory):
+    datadir = tmpdir_factory.mktemp('data')
+    filename = datadir + '/hdf5-test-3D.h5'
+    try:
+        with h5py.File(filename, 'r') as f:
+            yield f
+    except OSError:
+        with h5py.File(filename, "w") as f:
+            f.create_dataset("data", data=np.ones((17, 16, 16)))
+        with h5py.File(filename, 'r') as f:
+            yield f
+
+@pytest.fixture(scope='session')
 def random_hdf5(tmpdir_factory):
     datadir = tmpdir_factory.mktemp('data')
     filename = datadir + '/hdf5-test-random.h5'
@@ -84,6 +97,14 @@ def hdf5_ds_2(random_hdf5):
     ds = ds.initialize(InlineJobExecutor())
     return ds
 
+
+@pytest.fixture
+def hdf5_ds_3D(hdf5_3D):
+    ds = H5DataSet(
+        path=hdf5_3D.filename, ds_path="data", tileshape=(1, 16, 16)
+    )
+    ds = ds.initialize(InlineJobExecutor())
+    return ds
 
 @pytest.fixture
 def ds_complex():
