### 2020-01-30 8b0f3b3
Feedback; Add Alex as co-presenter
Thx @magnunor!

Not sure yet what exactly will be presented, but likely results from the planned hackathon in Berlin.
For that reason @sk1p as presenter as well
