### 2020-01-31 e419a93
Wording improvement
Ported from 86ece012a57231c41b065b059cc2a95e9d4a7139
