### 2020-02-03 60e3601
Peak finding correlation pattern in separate function
This allows users to get the correlation pattern for debugging and
visualization purposes.
