### 2020-02-03 91a20c8
Changelog entry for libertem.udf.blobfinder.correlation.get_correlation()
Forgot to git add...
