+++ docs/source/changelog/features/blobfinder_corr.rst
@@ -1,0 +1,4 @@
+[Feature] Access to peak finding correlation pattern
+====================================================
+
+* Give access to the correlation pattern for peak finding through :meth:`blb.correlation.get_correlation` (:pr:`571`)
