### 2020-02-03 458c2ea
Versionadded for new get_correlation()
