### 2020-02-04 e34b0c8
Clarify that the UDF produces raw results and not variance
run_stddev() is recommended instead. :-)
