### 2020-02-04 852d36b
Convenience: add AnalysisResult.__array__
This means we can call np.array(some_result) instead of
having to remember to use some_result.raw_data.
