### 2020-02-04 ab1ffcb
HDF5: fix reading of nD DataSets
 * Remove dimensionality assumptions in H5DataSet
 * Rewrite partition shape function to be more general
 * Improve performance of reading with roi. Not optimal, but
   should be much much better in case of sparse roi
