### 2020-02-04 c126568
make codeclimate happier with test function names
