### 2020-02-04 2bcbd4f
make codeclimate happier with test function names
