### 2020-02-04 ee1bc05
Fix test stability issue - need to sort files by name
