### 2020-02-04 ba19df0
DataSetMeta: add dtype (in addition to raw_dtype)
This is needed if, for some reason, the final output of the DataSet
won't match the on-disk dtype, for example if some preprocessing is
done in the process of reading. This fixes a FRMS6 test failure, which
surfaced because of the new picking udf.
