### 2020-02-04 LiberTEM/LiberTEM#554 @uellue
Results from some tests: UDFs can benefit significantly from optimizing the tile size and shape on a case-by-case basis. ApplyMasksUDF is particularly sensitive in case of deep mask stacks. In https://github.com/LiberTEM/LiberTEM-Examples/blob/master/bullseye-poly.ipynb, going from (1, 4, 250, 250) (1 MB) to (1, 16, 16, 250) (256 kB) improves the feature vector calculation performance by a factor of 2, while the stddev and logsum UDFs are much less affected.

That means tile size and shape negotiation can yield sizable performance gains.
