### 2020-02-05 15e0d8d
Detailed description for Job API deprecation
