+++ src/libertem/analysis/raw.py
@@ -17,6 +17,9 @@ class PickResultSet(AnalysisResultSet):
 
     .. versionadded:: 0.3.0
 
+    .. versionchanged:: 0.4.0.dev0
+        Picking now returns data in the native dtype of the dataset with the new UDF back-end.
+
     Attributes
     ----------
     intensity : libertem.analysis.base.AnalysisResult
