### 2020-02-05 fa98a4b
Port rawfft to UDF; add tests
* Use UDF back-end and disable JOB back-end
* Test for actual FFT
