### 2020-02-05 f203fdd
Deprecate Job API
* Trigger warning message when instantiating a Job
* Update documentation and docstrings
* Move Task from Job to UDF code base
* Detailed info in changelog
