+++ src/libertem/api.py
@@ -25,7 +25,7 @@ from libertem.udf.auto import AutoUDF
 class Context:
     """
     Context is the main entry point of the LiberTEM API. It contains
-    methods for loading datasets, creating jobs on them and running
+    methods for loading datasets, creating analyses on them and running
     them.
     """
 
@@ -100,6 +100,10 @@ class Context:
         Create a low-level mask application job. Each factory function should, when called,
         return a numpy array with the same shape as frames in the dataset (so dataset.shape.sig).
 
+        .. deprecated:: 0.4.0.dev0
+            Use :meth:`create_mask_analysis` or :class:`~libertem.udf.masks.ApplyMasksUDF`.
+            See also :ref:`job deprecation`.
+
         Parameters
         ----------
         factories : Union[Callable[[], array_like], Iterable[Callable[[], array_like]]]
@@ -160,21 +164,26 @@ class Context:
     def create_mask_analysis(self, factories, dataset, use_sparse=None,
                              mask_count=None, mask_dtype=None, dtype=None):
         """
-        Create a mask application analysis. Each factory function should, when called,
-        return a numpy array with the same shape as frames in the dataset (so dataset.shape.sig).
+        Create a mask application analysis. Each factory function should, when
+        called, return a numpy array with the same shape as frames in the
+        dataset (so dataset.shape.sig).
 
-        This is a more high-level method than `create_mask_job` and differs in the way the result
-        is returned. With `create_mask_job`, it is a single numpy array, here we split it up for
-        each mask we apply, make some default visualization available etc.
+        This is a more high-level interface than
+        :class:`~libertem.udf.masks.ApplyMasksUDF` and differs in the way the
+        result is returned. With :class:`~libertem.udf.masks.ApplyMasksUDF`, it
+        is a single numpy array, here we split it up for each mask we apply,
+        make some default visualization available etc.
 
         Parameters
         ----------
         factories : Union[Callable[[], array_like], Iterable[Callable[[], array_like]]]
-            Function or list of functions that take no arguments and create masks. The returned
-            masks can be numpy arrays, scipy.sparse or sparse https://sparse.pydata.org/ matrices.
-            The mask factories should not reference large objects because they can create
-            significant overheads when they are pickled and unpickled.
-            If a single function is specified, the first dimension is interpreted as the mask index.
+            Function or list of functions that take no arguments
+            and create masks. The returned masks can be numpy arrays,
+            scipy.sparse or sparse https://sparse.pydata.org/ matrices. The mask
+            factories should not reference large objects because they can create
+            significant overheads when they are pickled and unpickled. If a
+            single function is specified, the first dimension is interpreted as
+            the mask index.
         dataset : libertem.io.dataset.base.DataSet
             dataset to work on
         use_sparse : bool or None
@@ -184,17 +193,19 @@ class Context:
             * True: Convert all masks to sparse matrices.
             * False: Convert all masks to dense matrices.
         mask_count : int, optional
-            Specify the number of masks if a single factory function is used so that the
-            number of masks can be determined without calling the factory function.
+            Specify the number of masks if a single factory function is
+            used so that the number of masks can be determined without calling
+            the factory function.
         mask_dtype : numpy.dtype, optional
-            Specify the dtype of the masks so that mask dtype
-            can be determined without calling the mask factory functions. This can be used to
-            override the mask dtype in the result dtype determination. As an example, setting
-            this to np.float32 means that masks of type float64 will not switch the calculation
[…]
