+++ src/libertem/udf/FEM.py
@@ -74,26 +74,20 @@ def run_fem(ctx, dataset, center, rad_in, rad_out, roi=None):
     Parameters
     ----------
 
-    ctx: Context
-        Context class that contains methods for loading datasets,
-        creating jobs on them and running them
-
-    dataset: DataSet
+    ctx : libertem.api.Context
+    dataset : libertem.io.dataset.base.DataSet
         A dataset with 1- or 2-D scan dimensions and 2-D frame dimensions
-
-    center: tuple
+    center : tuple
         (x,y) - coordinates of a center of a ring for a masking region of interest to calculate SD
-
-    rad_in: int
+    rad_in : int
         Inner radius of a ring mask
-
-    rad_out: int
+    rad_out : int
         Outer radius of a ring mask
 
     Returns
     -------
 
-    pass_results: dict
+    pass_results : dict
         Returns a standard deviation(SD) value for each frame of pixels which belong to ring mask.
         To return 2-D array use pass_results['intensity'].data
 
