### 2020-02-06 f19cb82
remove stray hdfs extra from tox.ini
