+++ tox.ini
@@ -10,7 +10,6 @@ deps=
     -rtest_requirements.txt
 extras=
     hdbscan
-    hdfs
 setenv=
     # Using pytest in combination with tox on files that are part of the installed package
     # leads to collisions between the local source tree and the installed package when running tests.
