### 2020-02-06 7d02c7b
HDF5: automatically determine tileshape
