### 2020-02-06 bef42be
HDF5: automatically determine tileshape
