+++ docs/source/changelog/features/hdf5_tileshape_optional.rst
@@ -1,0 +1,5 @@
+[Feature] make tileshape parameter of HDF5 DataSet optional
+===========================================================
+
+ * Users should not worry about tileshape, so make the parameter
+   optional (for now kept for testing of non-standard tilings)
