### 2020-02-07 8f9c043
Review @sk1p, thx! :-)
* Link to classes and methods
* Deprecation info extended
* Use warnings module
* Tip on how to show DeprecationWarning in Jupyter etc
