+++ docs/source/changelog/misc/distributed.rst
@@ -1,0 +1,4 @@
+[Misc] Depend on newer version of Dask.distributed
+==================================================
+
+ * Depend on distributed>=2.2.0 because of an API change. (:pr:`577`)
