### 2020-02-07 ddc9bd0
Review @sk1p, thx! :-)
* Link to classes and methods
* Deprecation info extended
* Use warnings module
* Tip on how to show DeprecationWarning in Jupyter etc
