### 2020-02-07 7f79bfb
Info on restructuring in changelog
Based on #577
