+++ docs/source/changelog.rst
@@ -131,12 +131,12 @@ Restructuring the code base
 ---------------------------
 
 We are currently restructuring LiberTEM into modules that can be installed and
-used independently, see :issue:`261`. This will be a longer process and causes
-changes in the import locations. The first module to be separated is the
-`Blobfinder <https://github.com/LiberTEM/LiberTEM-blobfinder>_. For a transition
-period, importing from the previous locations is supported but will trigger a
-warning. See :ref:`show warnings` on how to activate deprecation warning
-messages, which is strongly recommended while the restructuring is ongoing.
+used independently, see :issue:`261`. This will be a longer process and changes
+the import locations. The first module to be separated is the `Blobfinder
+<https://github.com/LiberTEM/LiberTEM-blobfinder>`_. For a transition period,
+importing from the previous locations is supported but will trigger a warning.
+See :ref:`show warnings` on how to activate deprecation warning messages, which
+is strongly recommended while the restructuring is ongoing.
 
 .. _latest:
 .. _`v0-3-0`:
