### 2020-02-10 bb31d10
Fix issue 81
Open browser after starting the server.
cli has a new flag --no-browser , if present
the browser will not open.
