+++ src/libertem/web/server.py
@@ -2,6 +2,7 @@ import os
 import logging
 import asyncio
 import signal
+import webbrowser
 from functools import partial
 
 import tornado.web
@@ -115,13 +116,15 @@ def main(host, port, event_registry, shared_data):
     return app
 
 
-def run(host, port, local_directory):
+def run(host, port, no_browser, local_directory):
     # shared state:
     event_registry = EventRegistry()
     shared_data = SharedData()
 
     shared_data.set_local_directory(local_directory)
     main(host, port, event_registry, shared_data)
+    if not no_browser:
+        webbrowser.open(f'http://{host}:{port}')
     loop = asyncio.get_event_loop()
     signal.signal(signal.SIGINT, partial(sig_exit, shared_data=shared_data))
     # Strictly necessary only on Windows, but doesn't do harm in any case.
