+++ scripts/requirements.txt
@@ -1,0 +1,7 @@
+click
+pygithub
+requests
+twine
+html5lib
+docutils
+bibtexparser
