+++ .travis.yml
@@ -82,7 +82,7 @@ jobs:
       python: 3.7
       dist: xenial
       install:
-        - pip install click pygithub requests twine html5lib docutils bibtexparser
+        - pip install -r scripts/requirements.txt
       script:
         - python scripts/release upload --pypi-test-user=libertem_bot --pypi-user=libertem_bot --no-dry-run
 
