### 2020-02-11 784d0b5
Improved fix for issue 81
Avoid double negative implementation.
