+++ docs/source/changelog/features/open_browser.rst
@@ -1,0 +1,4 @@
+[Feature] Open browser after starting the server
+================================================
+
+ * Open browser by default, can be disabled using --no-browser (:issue:`81`)
