+++ docs/source/usage.rst
@@ -25,7 +25,7 @@ log output::
 
     [2018-08-08 13:57:58,266] INFO [libertem.web.server.main:886] listening on localhost:9000
 
-There are a few command line options available:: 
+There are a few command line options available::
 
     Usage: libertem-server [OPTIONS]
 
@@ -34,7 +34,7 @@ There are a few command line options available::
       --help          Show this message and exit.
 
 To access LiberTEM remotely, you can use :ref:`use SSH forwarding <ssh forwarding>`.
-    
+
 As there is currently no authentication yet, listening on a different host than
 `127.0.0.1` / `localhost` is disabled. As a workaround, if you want to access LiberTEM from a different computer,
 you can use ssh port forwarding (example command for conda):
@@ -51,15 +51,27 @@ Or, with virtualenv:
 
 This then makes LiberTEM that is running on `remote-hostname` available on your local host via http://localhost:9000/
 
+.. versionadded:: 0.4.0.dev0
+
+.. code-block:: shell
+
+    (libertem) $ libertem-server
+
+This will launch a new browser tab showing GUI by default.
+
+New command line option ::
+
+    --browser / --no-browser  this flag enable/disable opening the browser
+
 
 The user interface
 ------------------
 
 After starting the server, you can open the GUI in your browser. By default it will be at http://localhost:9000 .
-At the beginning, the GUI shows a prompt to create a local cluster or connect to a running one. The number of workers is preset with a number 
+At the beginning, the GUI shows a prompt to create a local cluster or connect to a running one. The number of workers is preset with a number
 that will likely give optimal performance on the given machine.
 
-..  figure:: ./images/use/create.png 
+..  figure:: ./images/use/create.png
 
 After connection to a cluster, LiberTEM shows a button to start browsing for available files. On a local cluster that's simply
 the local filesystem.
@@ -81,13 +93,13 @@ Once a dataset is loaded, you can add analyses to it. As an example we choose a 
 
 ..  figure:: ./images/use/add_analysis.png
 
-The GUI shows two windows: On the left it shows the current mask. Directly after adding the analysis, LiberTEM starts calculating an average of all the detector frames. As soon as this is finished, the average is overlaid with the mask to help the user with positioning the virtual detector. The window on the right will later show the result of applying the mask to the data. In the beginning it is empty. The first processing might take a while depending on file size and IO performance. Fast SSDs and enough RAM to keep the working files in the file system cache are highly recommended for a good user experience. 
+The GUI shows two windows: On the left it shows the current mask. Directly after adding the analysis, LiberTEM starts calculating an average of all the detector frames. As soon as this is finished, the average is overlaid with the mask to help the user with positioning the virtual detector. The window on the right will later show the result of applying the mask to the data. In the beginning it is empty. The first processing might take a while depending on file size and IO performance. Fast SSDs and enough RAM to keep the working files in the file system cache are highly recommended for a good user experience.
 
 You can adjust the virtual detector by dragging the handles in the GUI. Below it shows the parameters in numerical form. This is useful to extract positions, for example for scripting.
 
 ..  figure:: ./images/use/adjust.png
 
-After clicking "Apply", LiberTEM performs the calculation and shows the result in scan coordinates on the right. 
+After clicking "Apply", LiberTEM performs the calculation and shows the result in scan coordinates on the right.
 
 ..  figure:: ./images/use/apply.png
 
