### 2020-02-11 8ff5a3a
Improved fix for issue 81
Avoid double negative implementation.
