+++ src/libertem/web/cli.py
@@ -7,13 +7,13 @@ import os
               default=9000, type=int)
 @click.option('--local-directory', help='local directory to manage dask-worker-space files',
               default='dask-worker-space', type=str)
-@click.option('--no-browser', help='this flag disables opening the browser', is_flag=True)
+@click.option('--browser/--no-browser', help='this flag enable/disable opening the browser', default='True')
 # FIXME: the host parameter is currently disabled, as it poses a security risk
 # as long as there is no authentication
 # see also: https://github.com/LiberTEM/LiberTEM/issues/67
 # @click.option('--host', help='host on which the server should listen on',
 #               default="localhost", type=str)
-def main(port, local_directory, no_browser, host="localhost"):
+def main(port, local_directory, browser, host="localhost"):
     os.environ.setdefault('OMP_NUM_THREADS', '1')
     os.environ.setdefault('MKL_NUM_THREADS', '1')
     os.environ.setdefault('OPENBLAS_NUM_THREADS', '1')
@@ -21,4 +21,4 @@ def main(port, local_directory, no_browser, host="localhost"):
     from libertem.cli_tweaks import console_tweaks
     from .server import run
     console_tweaks()
-    run(host, port, no_browser, local_directory)
+    run(host, port, browser, local_directory)
