### 2020-02-12 40da4bf
Fix issue 81
flake 8 error resolved.
documentation improved.
