+++ src/libertem/web/cli.py
@@ -7,7 +7,7 @@ import os
               default=9000, type=int)
 @click.option('--local-directory', help='local directory to manage dask-worker-space files',
               default='dask-worker-space', type=str)
-@click.option('--browser/--no-browser', help='this flag enable/disable opening the browser', default='True')
+@click.option('--browser/--no-browser', help='enable/disable opening the browser', default='True')
 # FIXME: the host parameter is currently disabled, as it poses a security risk
 # as long as there is no authentication
 # see also: https://github.com/LiberTEM/LiberTEM/issues/67
