### 2020-02-12 e96b670
Replace blobfinder with imports from libertem_blobfinder
Still TODO:
* setup.py
* Documentation
