+++ src/libertem/udf/blobfinder/correlation.py
@@ -1,641 +1,14 @@
-import functools
-import os
-
-import numpy as np
-from skimage.feature import peak_local_max
-import numba
-
-from libertem.udf import UDF
-import libertem.masks as masks
-from libertem.common.container import MaskContainer
-
-from .patterns import MatchPattern
-
-# FIXME There's work on flexible FFT backends in scipy
-# https://github.com/scipy/scipy/wiki/GSoC-2019-project-ideas#revamp-scipyfftpack
-# and discussions about pyfftw performance vs other implementations
-# https://github.com/pyFFTW/pyFFTW/issues/264
-# For that reason we shoud review the state of Python FFT implementations
-# regularly and adapt our choices accordingly
-try:
-    import pyfftw
-    fft = pyfftw.interfaces.numpy_fft
-    pyfftw.interfaces.cache.enable()
-    zeros = pyfftw.zeros_aligned
-except ImportError:
-    fft = np.fft
-    zeros = np.zeros
-
-# Necessary to work with JIT disabled for coverage and testing purposes
-# https://github.com/LiberTEM/LiberTEM/issues/539
-if os.getenv('NUMBA_DISABLE_JIT'):
-    def to_fixed_tuple(arr, l):
-        return tuple(arr)
-else:
-    from numba.unsafe.ndarray import to_fixed_tuple
-
-
-def get_correlation(sum_result, match_pattern: MatchPattern):
-    '''
-    Calculate the correlation between :code:`sum_result` and :code:`match_pattern`.
-
-    .. versionadded:: 0.4.0.dev0
-
-    Parameters
-    ----------
-
-    sum_result: numpy.ndarray
-        2D result frame as correlation input
-    match_pattern : MatchPattern
-        Instance of :class:`~libertem.udf.blobfinder.MatchPattern` to correlate
-        :code:`sum_result` with
-    '''
-    spec_mask = match_pattern.get_template(sig_shape=sum_result.shape)
-    spec_sum = fft.rfft2(sum_result)
-    corrspec = spec_mask * spec_sum
-    return fft.fftshift(fft.irfft2(corrspec))
-
-
-def get_peaks(sum_result, match_pattern: MatchPattern, num_peaks):
-    '''
-    Find peaks of the correlation between :code:`sum_result` and :code:`match_pattern`.
-
-    The result  can then be used as input to
-    :meth:`~libertem.analysis.fullmatch.FullMatcher.full_match`
-    to extract grid parameters, :meth:`~libertem.udf.blobfinder.correlation.run_fastcorrelation`
-    to find the position in each frame or to construct a mask to extract feature vectors with
-    :meth:`~libertem.udf.blobfinder.utils.feature_vector`.
-
-    Parameters
-    ----------
-
-    sum_result: numpy.ndarray
-        2D result frame as correlation input
-    match_pattern : MatchPattern
-        Instance of :class:`~libertem.udf.blobfinder.MatchPattern` to correlate
-        :code:`sum_result` with
-    num_peaks : int
-        Number of peaks to find
-
-    Example
-    -------
-
-    >>> frame, _, _ = libertem.utils.generate.cbed_frame(radius=4)
-    >>> pattern = libertem.udf.blobfinder.RadialGradient(radius=4)
-    >>> peaks = get_peaks(frame[0], pattern, 7)
-    >>> print(peaks)
-    [[64 64]
-     [64 80]
-     [80 80]
-     [80 64]
-     [48 80]
-     [48 64]
-     [64 96]]
-    '''
-    corr = get_correlation(sum_result, match_pattern)
-    peaks = peak_local_max(corr, num_peaks=num_peaks)
-    return peaks
[…]
