+++ src/libertem/udf/blobfinder/utils.py
@@ -1,129 +1,13 @@
-import numpy as np
-import matplotlib.pyplot as plt
+import warnings
 
-import libertem.masks as masks
-import libertem.analysis.gridmatching as grm
-
-from .patterns import MatchPattern
-
-# FIXME There's work on flexible FFT backends in scipy
-# https://github.com/scipy/scipy/wiki/GSoC-2019-project-ideas#revamp-scipyfftpack
-# and discussions about pyfftw performance vs other implementations
-# https://github.com/pyFFTW/pyFFTW/issues/264
-# For that reason we shoud review the state of Python FFT implementations
-# regularly and adapt our choices accordingly
-try:
-    import pyfftw
-    fft = pyfftw.interfaces.numpy_fft
-    pyfftw.interfaces.cache.enable()
-    zeros = pyfftw.zeros_aligned
-except ImportError:
-    fft = np.fft
-    zeros = np.zeros
+from libertem_blobfinder.common.patterns import feature_vector  # noqa: F401
+from libertem_blobfinder.udf.utils import visualize_frame, paint_markers  # noqa: F401
 
 
-def feature_vector(imageSizeX, imageSizeY, peaks, match_pattern: MatchPattern):
-    '''
-    This function generates a sparse mask stack to extract a feature vector.
-
-    A match template based on the parameters in :code:`parameters` is placed at
-    each peak position in an individual mask layer. This mask stack can then
-    be used with :class:`~libertem.udf.masks.ApplyMasksUDF` to generate a feature vector for each
-    frame.
-
-    Summing up the mask stack along the first axis generates a mask that can be used for virtual
-    darkfield imaging of all peaks together.
-
-    Parameters
-    ----------
-
-    imageSizeX,imageSizeY : int
-        Frame size in px
-    peaks : numpy.ndarray
-        Peak positions in px as numpy.ndarray of shape (n, 2) with integer type
-    match_pattern : MatchPattern
-        Instance of :class:`~MatchPattern`
-    '''
-    crop_size = match_pattern.get_crop_size()
-    return masks.sparse_template_multi_stack(
-        mask_index=range(len(peaks)),
-        offsetX=peaks[:, 1] - crop_size,
-        offsetY=peaks[:, 0] - crop_size,
-        template=match_pattern.get_mask((2*crop_size + 1, 2*crop_size + 1)),
-        imageSizeX=imageSizeX,
-        imageSizeY=imageSizeY,
-    )
-
-
-def visualize_frame(ctx, ds, result, indices, r, y, x, axes, colors=None, stretch=10):
-    '''
-    Visualize the refinement of a specific frame in matplotlib axes
-    '''
-    # Get the frame from the dataset
-    get_sample_frame = ctx.create_pick_analysis(dataset=ds, y=y, x=x)
-    sample_frame = ctx.run(get_sample_frame)
-
-    d = sample_frame[0].raw_data
-
-    pcm = axes.imshow(np.log(d - np.min(d) + 1))
-
-    refined = result['refineds'].data[y, x]
-    elevations = result['peak_elevations'].data[y, x]
-    selector = result['selector'].data[y, x]
-
-    max_elevation = np.max(elevations)
-
-    # Calclate the best fit positions to compare with the
-    # individual peak positions.
-    # A difference between best fit and individual peaks highlights outliers.
-    calculated = grm.calc_coords(
-        zero=result['zero'].data[y, x],
-        a=result['a'].data[y, x],
-        b=result['b'].data[y, x],
-        indices=indices
-    )
-
-    paint_markers(
-        axes=axes,
-        r=r,
-        refined=refined,
-        normalized_elevations=elevations/max_elevation,
-        calculated=calculated,
[…]

[…]
