### 2020-02-12 750e3a7
Add versioning note and bump to .dev0 step
Also fixes rc version syntax (no dot before rcN!)
