### 2020-02-12 da43218
Test against RC of libertem-blobfinder
