### 2020-02-12 768252b
Test against RC of libertem-blobfinder
