+++ tox.ini
@@ -8,6 +8,8 @@ commands=
     pytest --doctest-modules --ignore=src/libertem/win_tweaks.py src/libertem/
 deps=
     -rtest_requirements.txt
+    # Inject release candidates for testing
+    -roverride_requirements.txt
 extras=
     hdbscan
 setenv=
