### 2020-02-12 acda32f
Documentation update for blobfinder separation
TODO: Examples
