### 2020-02-12 e940c77
Documentation update for blobfinder separation
TODO: Examples
