+++ docs/source/conf.py
@@ -205,8 +205,9 @@ doctest_test_doctest_blocks = ''
 todo_include_todos = True
 
 intersphinx_mapping = {
-    'numpy': ('http://docs.scipy.org/doc/numpy/', None),
-    'python': ('https://docs.python.org/3.7', None)
+    'numpy': ('https://docs.scipy.org/doc/numpy/', None),
+    'python': ('https://docs.python.org/3.7', None),
+    'libertem_blobfinder': ('https://libertem.github.io/LiberTEM-blobfinder/', None),
 }
 
 # Sphinx' link checker.
