+++ docs/source/changelog/bugfix/blobfinder.rst
@@ -1,5 +1,0 @@
-[Bugfix] Bounds checking in blobfinder
-======================================
-
- * Fix bounds checking, size and index calculation bugs in blobfinder (:issue:`539`, :pr:`548`)
- 
