+++ docs/source/packages.rst
@@ -1,0 +1,14 @@
+.. _packages:
+
+Package overview
+================
+
+LiberTEM is currently being split into several application-specific sub-packages
+to keep dependencies manageable and facilitate using LiberTEM code in other
+projects and vice-versa. See also :issue:`261`.
+
+This list provides an overview of the current sub-packages and is constantly updated.
+
+:libertem-blobfinder:
+    Since 0.4.0.dev0: Correlation-based peak finding and strain mapping.
+    https://libertem.github.io/LiberTEM-blobfinder/
