+++ override_requirements.txt
@@ -1,3 +1,3 @@
 # Inject release candidates for LiberTEM sub-packages during release and testing process
 --index-url https://test.pypi.org/simple/ --extra-index-url https://pypi.org/simple
-libertem-blobfinder
+# libertem-blobfinder
