+++ tox.ini
@@ -77,6 +77,8 @@ deps=
     nbsphinx
     nbsphinx_link
     ipython
+    # Inject release candidates for doctest testing
+    -roverride_requirements.txt
 skipsdist=True
 
 [testenv:docs-deploy]
@@ -96,6 +98,8 @@ deps=
     nbsphinx
     nbsphinx_link
     ipython
+    # Inject release candidates for doctest testing
+    -roverride_requirements.txt
 skipsdist=True
 passenv=
     HOME
@@ -120,6 +124,8 @@ deps=
     nbsphinx
     nbsphinx_link
     ipython
+    # Inject release candidates for doctest testing
+    -roverride_requirements.txt
 skipsdist=True
 whitelist_externals=
     cat
