+++ src/libertem/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "0.4.0.dev0"
+__version__ = "0.4.0rc0"
