+++ docs/source/app/strain.rst
@@ -3,6 +3,6 @@
 Strain mapping
 ==============
 
-.. versionchanged:: 0.4.0.dev0
+.. versionchanged:: 0.4.0
     Blobfinder has been spun out as a separate package. Documentation is
     available at https://libertem.github.io/LiberTEM-blobfinder/
