+++ docs/source/applications.rst
@@ -4,7 +4,7 @@ Applications
 ============
 
 .. note::
-    Starting from 0.4.0.dev0, the application-specific code of LiberTEM will be
+    Starting from 0.4.0, the application-specific code of LiberTEM will be
     spun out step by step as sub-packages that can be installed independent of
     LiberTEM core. See :ref:`packages` for the current overview of sub-packages.
 
