+++ docs/source/packages.rst
@@ -10,5 +10,5 @@ projects and vice-versa. See also :issue:`261`.
 This list provides an overview of the current sub-packages and is constantly updated.
 
 :libertem-blobfinder:
-    Since 0.4.0.dev0: Correlation-based peak finding and strain mapping.
+    Since 0.4.0: Correlation-based peak finding and strain mapping.
     https://libertem.github.io/LiberTEM-blobfinder/
