+++ docs/source/changelog/misc/udfmasks.rst
@@ -1,4 +1,0 @@
-[Misc] UDF interface for analyses
-=================================
-
- * All analyses ported from Job to UDF back-end. The Job-related code remains for now for comparison purposes (:issue:`549`, :pr:`550`)
