### 2020-02-13 6a22c4d
Documentation updates
* Include MemoryDataSet
* Rewrap
* Various smaller touch-ups
