+++ README.rst
@@ -62,7 +62,7 @@ Applications
 - Virtual detectors (virtual bright field, virtual HAADF, center of mass :cite:`Krajnak2016`,
   custom shapes via masks)
 - `Analysis of amorphous materials <https://libertem.github.io/LiberTEM/app/amorphous.html>`_
-- `Strain mapping <https://libertem.github.io/LiberTEM/app/strain.html>`_
+- `Strain mapping <https://libertem.github.io/LiberTEM-blobfinder/>`_
 - `Custom analysis functions (user-defined functions) <https://libertem.github.io/LiberTEM/udf.html>`_
 - `Off-axis electron holography reconstruction <https://libertem.github.io/LiberTEM/app/holography.html>`_
 
