+++ docs/source/usage.rst
@@ -38,13 +38,14 @@ There are a few command line options available::
       --help          Show this message and exit.
 
 .. versionadded:: 0.4.0
-    --browser / --no-browser option is added, open browser by default. 
+    :code:`--browser` / :code:`--no-browser` option is added, open browser by default. 
 
 To access LiberTEM remotely, you can use :ref:`use SSH forwarding <ssh forwarding>`.
 
 As there is currently no authentication yet, listening on a different host than
-`127.0.0.1` / `localhost` is disabled. As a workaround, if you want to access LiberTEM from a different computer,
-you can use ssh port forwarding (example command for conda):
+:code:`127.0.0.1` / :code:`localhost` is disabled. As a workaround, if you want
+to access LiberTEM from a different computer, you can use ssh port forwarding
+(example command for conda):
 
 .. code-block:: shell
 
@@ -56,68 +57,100 @@ Or, with virtualenv:
 
      $ ssh -L 9000:localhost:9000 <remote-hostname> "/path/to/virtualenv/bin/libertem-server"
 
-This then makes LiberTEM that is running on `remote-hostname` available on your local host via http://localhost:9000/
+This then makes LiberTEM that is running on :code:`remote-hostname` available on
+your local host via http://localhost:9000/
 
 
 The user interface
 ------------------
 
-After starting the server, you can open the GUI in your browser. By default it will be at http://localhost:9000 .
-At the beginning, the GUI shows a prompt to create a local cluster or connect to a running one. The number of workers is preset with a number
-that will likely give optimal performance on the given machine.
+After starting the server, you can open the GUI in your browser. By default it
+will be at http://localhost:9000 . At the beginning, the GUI shows a prompt to
+create a local cluster or connect to a running one. The number of workers is
+preset with a number that will likely give optimal performance on the given
+machine.
 
 ..  figure:: ./images/use/create.png
 
-After connection to a cluster, LiberTEM shows a button to start browsing for available files. On a local cluster that's simply
-the local filesystem.
+After connection to a cluster, LiberTEM shows a button to start browsing for
+available files. On a local cluster that's simply the local filesystem.
 
 ..  figure:: ./images/use/browse.png
 
-This opens the file browser dialogue. On top it shows the current directory, below it lists all files and subdirectories in that directory. You select an entry by clicking once on it.
-You can move up one directory with the ".." entry on top of the list. The file browser is still very basic. Possible improvements are discussed in `Issue #83 <https://github.com/LiberTEM/LiberTEM/issues/83>`_. Contributions are highly appreciated!
+This opens the file browser dialogue. On top it shows the current directory,
+below it lists all files and subdirectories in that directory. You select an
+entry by clicking once on it. You can move up one directory with the ".." entry
+on top of the list. The file browser is still very basic. Possible improvements
+are discussed in `Issue #83 <https://github.com/LiberTEM/LiberTEM/issues/83>`_.
+Contributions are highly appreciated!
 
 ..  figure:: ./images/use/open.png
 
-After selecting a file, you set the type in the drop-down menu at the top of the dialogue above the file name. After that you set the appropriate parameters that depend on the file type. Clicking on "Load Dataset" will open the file with the selected parameters. The interface and internal logic to find good presets based on file type and available metadata, validate the inputs and display helpful error messages is still work in progress. Contributions are highly appreciated!
+After selecting a file, you set the type in the drop-down menu at the top of the
+dialogue above the file name. After that you set the appropriate parameters that
+depend on the file type. Clicking on "Load Dataset" will open the file with the
+selected parameters. The interface and internal logic to find good presets based
+on file type and available metadata, validate the inputs and display helpful
[…]
