### 2020-02-13 e1277b2
Documentation updates
* Include MemoryDataSet
* Rewrap
* Various smaller touch-ups
