### 2020-02-13 5fa9254
Improved docs for installation
added steps for installing LiberTEM for code contribution.
