+++ docs/source/install.rst
@@ -14,13 +14,13 @@ Installation
     clone`_ use pip, but they do fundamentally different things. :code:`pip
     install libertem` downloads the latest release from PyPi, which can be
     somewhat older.
-    
+
     Changing directory to a git clone and running :code:`pip install -e .`
     installs from the local directory in editable mode. "Editable mode" means
     that the source directory is linked into the current Python environment
     rather than copied. That means changes in the source directory are
     immediately active in the Python environment.
-    
+
     Installing from a git clone in editable mode is the correct setup for
     development work and using :ref:`the latest features in the development
     branch <continuous>`. Installing from PyPI is easier and preferred for new
@@ -64,7 +64,7 @@ Replace :code:`~/libertem-venv/` with any path where you would like to create
 the venv. You can then activate the virtualenv with
 
 .. code-block:: shell
-    
+
     $ source ~/libertem-venv/bin/activate
 
 Afterwards, your shell prompt should be prefixed with :code:`(libertem)` to
@@ -72,7 +72,7 @@ indicate that the environment is active:
 
 .. code-block:: shell
 
-    (libertem) $ 
+    (libertem) $
 
 For more information about virtualenv, for example if you are using a shell
 without `source`, please `refer to the virtualenv documentation
@@ -100,7 +100,7 @@ command:
 To install or later run LiberTEM, activate the environment with the following
 command:
 
-.. code-block:: shell  
+.. code-block:: shell
 
     $ source activate libertem
 
@@ -109,10 +109,10 @@ indicate that the environment is active:
 
 .. code-block:: shell
 
-    (libertem) $ 
+    (libertem) $
 
 Now the environment is ready to install LiberTEM.
-    
+
 For more information about conda, see their `documentation about creating and
 managing environments
 <https://conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html>`_.
@@ -135,12 +135,28 @@ continue by reading the :ref:`usage documentation`.
 Installing from a git clone
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-If you want to follow the latest development or contribute to LiberTEM, you
-should install LiberTEM from a git clone:
+If you want to follow the latest development, you should install LiberTEM from
+a git clone:
 
 .. code-block:: shell
 
     $ git clone https://github.com/LiberTEM/LiberTEM
+
+Or if you wish to contribute to LiberTEM, follow these steps instead :
+
+#. Log into your `GitHub <https://github.com/>`_ account.
+
+#. Go to the `LiberTEM GitHub <https://github.com/liberteM/LiberTEM/>`_ home page.
+
+#. Click on the *fork* button:
+
+    .. image:: ../images/forking_button.png #require change here
+
+#. Copy your fork of LiberTEM from GitHub to your computer
+
+.. code-block:: shell
+
+    $ git clone https://github.com/your-user-name/LiberTEM
 
 Activate the Python environment (conda or virtualenv) and change to the newly
 created directory with the clone of the LiberTEM repository. Now you can start
@@ -148,7 +164,7 @@ the LiberTEM installation. Please note the dot at the end, which indicates the
 current directory!
 
 .. code-block:: shell
-    
+
     (libertem) $ pip install -e .
 
 This should download the dependencies and install LiberTEM in the environment.
@@ -183,6 +199,10 @@ The full grid matching routines in :py:mod:`libertem.analysis.fullmatch` depend
 on `HDBSCAN <https://hdbscan.readthedocs.io/en/latest/>`_. This is an optional
 dependency because of installation issues on some platforms.
 
+The fast correlation methods in :py:mod:`libertem.udf.blobfinder` can optionally
[…]
