### 2020-02-13 8cb5f18
Improved docs for installation
added steps for installing LiberTEM for code contribution.
