### 2020-02-13 f2a8b84
Improve docs regarding installation
minor fix
