### 2020-02-13 26090cc
Improve docs regarding installation
minor fix
