+++ docs/source/install.rst
@@ -150,7 +150,7 @@ Or if you wish to contribute to LiberTEM, follow these steps instead :
 
 #. Click on the *fork* button:
 
-    .. image:: ../images/forking_button.png #require change here
+    .. image:: ../images/forking_button.png
 
 #. Copy your fork of LiberTEM from GitHub to your computer
 
@@ -198,10 +198,6 @@ Other extra packages
 The full grid matching routines in :py:mod:`libertem.analysis.fullmatch` depend
 on `HDBSCAN <https://hdbscan.readthedocs.io/en/latest/>`_. This is an optional
 dependency because of installation issues on some platforms.
-
-The fast correlation methods in :py:mod:`libertem.udf.blobfinder` can optionally
-use `pyFFTW <https://hgomersall.github.io/pyFFTW/>`_ instead of the built-in
-Numpy FFT.
 
 Updating
 ~~~~~~~~
