### 2020-02-13 7b0cca5
Minor fix on docs
added the line that loads image, removed the one
that wasn't working.
