### 2020-02-13 8ff2f93
bump version: 0.4.0rc1 → 0.4.0
