### 2020-02-14 550d391
Added link to github documentation
Link to fork a repo documentation by github is added.
replaced copy with clone.
