### 2020-02-17 f457579
HDF5: fix tile offset if reading with roi
The roi needs to be flat when calculating the frame offset.
