### 2020-02-17 243899e
Add changelog entry for #606
