+++ docs/source/changelog/bugfix/hdf5_roi_fix.rst
@@ -1,0 +1,4 @@
+[Bugfix] Fix reading from HDF5 with roi
+=======================================
+
+* Offset calculation for tiles was wrong beyond the first partition (:pr:`606`)
