### 2020-02-17 cd618f7
Add changelog entry for #606
