+++ docs/source/changelog/misc/removeme.rst
@@ -1,4 +1,0 @@
-No changes yet
-==============
-
-No changes yet. Remove this file as soon as the first real change log entry is created.
