### 2020-02-18 f794e88
Move DataSet initialization into `load` function
