+++ src/libertem/io/dataset/__init__.py
@@ -1,4 +1,5 @@
 import importlib
+
 from libertem.io.dataset.base import DataSetException
 
 
@@ -16,19 +17,26 @@ filetypes = {
 }
 
 
-def load(filetype, *args, **kwargs):
+def load(filetype, executor, *args, **kwargs):
     """
-    load a dataset
+    Low-level method to load a dataset. Usually you will want
+    to use Context.load instead!
 
     Parameters
     ----------
     filetype : str or DataSet type
         see libertem.io.dataset.filetypes for supported types, example: 'hdf5'
 
+    executor : JobExecutor
+
     additional parameters are passed to the concrete DataSet implementation
     """
     cls = get_dataset_cls(filetype)
-    return cls(*args, **kwargs)
+    ds = cls(*args, **kwargs)
+    ds = ds.initialize(executor)
+    ds.set_num_cores(len(executor.get_available_workers()))
+    executor.run_function(ds.check_valid)
+    return ds
 
 
 def register_dataset_cls(filetype, cls):
@@ -59,9 +67,7 @@ def get_dataset_cls(filetype):
     return cls
 
 
-def detect(path, executor=None):
-    if executor is None:
-        raise TypeError("FIXME")
+def detect(path, executor):
     for filetype in filetypes.keys():
         try:
             cls = get_dataset_cls(filetype)
