### 2020-02-18 ff17ef0
Documentation QA
Small updates
