### 2020-02-18 513f615
Documentation QA
Small updates
