### 2020-02-18 48e418e
load: actually support async executor etc.
