+++ src/libertem/utils/async_utils.py
@@ -1,0 +1,10 @@
+from functools import partial
+
+import tornado
+
+
+async def run_blocking(fn, *args, **kwargs):
+    """
+    run blocking function fn with args, kwargs in a thread and return a corresponding future
+    """
+    return await tornado.ioloop.IOLoop.current().run_in_executor(None, partial(fn, *args, **kwargs))
