+++ src/libertem/analysis/clust.py
@@ -1,6 +1,6 @@
 from libertem.viz import visualize_simple
 from .base import BaseAnalysis, AnalysisResult, AnalysisResultSet
-from libertem.web.base import run_blocking
+from libertem.utils.async_utils import run_blocking
 from libertem.executor.base import JobCancelledError
 import libertem.udf.feature_vector_maker as feature
 from libertem.udf.base import UDFRunner
