### 2020-02-18 d2ddd64
load: actually support async executor etc.
