### 2020-02-18 a6c446b
HDF5: Fix crash w/ full_frame + no tileshape param
