### 2020-02-18 fb0b78a
HDF5: Fix crash w/ full_frame + no tileshape param
