+++ docs/source/changelog/bugfix/hdf5_full_frame_auto_tileshape.rst
@@ -1,0 +1,4 @@
+[Bugfix] Fix HDF5 with automatic tileshape 
+==========================================
+
+ * Fix type confusion error: tuple vs :code:`Shape` (:pr:`608`)
