### 2020-02-18 2eeab88
Add `get_extensions` function
Returns a set of supported file extensions for all datasets
