### 2020-02-18 a9e5c95
Add `get_extensions` function
Returns a set of supported file extensions for all datasets
