### 2020-02-18 c18e2d4
PickUDF now complains about all-zero ROI
Fixes #609
