+++ tests/udf/test_pick.py
@@ -1,3 +1,4 @@
+import pytest
 import numpy as np
 
 from libertem.udf.raw import PickUDF
@@ -17,3 +18,16 @@ def test_pick(lt_ctx):
 
     assert np.allclose(data[roi], res['intensity'].data)
     assert data.dtype == res['intensity'].data.dtype
+
+
+def test_pick_zero_roi(lt_ctx):
+    data = _mk_random(size=(16, 16, 16, 16), dtype="float32")
+    dataset = MemoryDataSet(data=data, tileshape=(3, 7, 7),
+                            num_partitions=7, sig_dims=2)
+    roi = np.zeros(dataset.shape.nav, dtype=bool)
+
+    udf = PickUDF()
+    with pytest.raises(ValueError) as e:
+        res = lt_ctx.run_udf(dataset=dataset, udf=udf, roi=roi)
+
+    assert e.match("cannot load data with all-zeros ROI")
