### 2020-02-18 ed75a73
Add a progress bar with tqdm, closes #612
