### 2020-02-18 eb98bc5
Add a progress bar with tqdm, closes #612
