+++ setup.py
@@ -156,6 +156,7 @@ setup(
         'jsonschema',
         'scikit-learn',
         'msgpack<1',  # FIXME: remove after distributed is fixed
+        'tqdm',
         # FIXME remove in 0.7.0 after blobfinder deprecation
         'libertem-blobfinder',
     ],
