+++ docs/source/changelog/features/progress_bar.rst
@@ -1,0 +1,6 @@
+[Feature] Add an optional progress bar
+======================================
+
+* Can be enabled by passing :code:`progress=True` to the
+  :meth:`~libertem.api.Context.run_udf` method.
+* Add a dependency on tqdm
