### 2020-02-18 e4c96fd
Add PR reference to changelog
