+++ docs/source/changelog/features/progress_bar.rst
@@ -2,5 +2,5 @@
 ======================================
 
 * Can be enabled by passing :code:`progress=True` to the
-  :meth:`~libertem.api.Context.run_udf` method.
+  :meth:`~libertem.api.Context.run_udf` method (:pr:`613`).
 * Add a dependency on tqdm
