### 2020-02-18 eff1034
Move SumUDF to UDF folder and improve documentation
