+++ src/libertem/udf/raw.py
@@ -16,6 +16,10 @@ class PickUDF(UDF):
 
     .. versionadded:: 0.4.0
     '''
+    def __init__(self):
+        ''
+        super().__init__()
+
     def get_preferred_input_dtype(self):
         ''
         return self.USE_NATIVE_DTYPE
