### 2020-02-18 ccca7c9
Move SumUDF to UDF folder and improve documentation
