### 2020-02-18 8b5f2fd
add progressbar smoke test
