+++ src/libertem/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "0.4.0"
+__version__ = "0.4.1rc0"
