+++ docs/source/changelog/bugfix/clustering.rst
@@ -1,0 +1,6 @@
+[Bugfix] Performance improvements of clustering analysis
+========================================================
+
+* Use a connectivity matrix to only cluster neighboring pixels,
+  reducing memory footprint while improving speed and quality (:pr:`617`)
+* Use faster :class:`~libertem.udf.masks.ApplyMasksUDF` to generate feature vector (:pr:`617`)
