### 2020-02-18 4121817
Improve clustering analysis
* Reduce memory footprint and improve speed by using a connectivity matrix
* Use faster ApplyMasksUDF to generate a feature vector
* Don't use delta and binary vector, instead use normalized float values

TODO update screenshots and documentation
