+++ docs/source/changelog/features/get_extensions.rst
@@ -1,4 +1,4 @@
-[Feature] `:code:`libertem.io.dataset.get_extensions`
-=====================================================
+[Feature] :code:`libertem.io.dataset.get_extensions`
+====================================================
 
 * Returns a set of supported file extensions for all datasets
