### 2020-02-18 b0beb21
Cleanup for changelog items
 * Remove items that were released as part of v0.4.1
 * Fix typo/syntax error in headline of get_extensions.rst
