### 2020-02-18 73c5350
Update figures and docs; adjust default number of pixels
