+++ docs/source/changelog/bugfix/clustering.rst
@@ -2,5 +2,5 @@
 ========================================================
 
 * Use a connectivity matrix to only cluster neighboring pixels,
-  reducing memory footprint while improving speed and quality (:pr:`617`)
-* Use faster :class:`~libertem.udf.masks.ApplyMasksUDF` to generate feature vector (:pr:`617`)
+  reducing memory footprint while improving speed and quality (:pr:`618`)
+* Use faster :class:`~libertem.udf.masks.ApplyMasksUDF` to generate feature vector (:pr:`618`)
