### 2020-02-18 4f7c34b
Update figures and docs; adjust default number of pixels
