### 2020-02-18 dc83042
run_for_dataset: try to reduce complexity a bit
