### 2020-02-18 a30ece5
Flake8 fix
