### 2020-02-18 dd725d7
move precondition check into its own method
