### 2020-02-18 63628f4
Deprecate FeatureVecMakerUDF
