+++ docs/source/changelog/obsolescence/featurevector.rst
@@ -1,0 +1,8 @@
+[Deprecation] FeatureVecMakerUDF
+================================
+
+* :class:`libertem.udf.feature_vector_maker.FeatureVecMakerUDF` is deprecated
+  and will be removed in 0.6.0. Use :class:`~libertem.udf.masks.ApplyMasksUDF`
+  with a sparse stack of single pixel masks or a stack generated by
+  :meth:`libertem_blobfinder.common.patterns.feature_vector` instead.
+  (:pr:`618`)
