### 2020-02-18 b0d3717
Updated more information
added more resources for git practice.
