### 2020-02-19 019e126
Add Context.dataset_info method
