+++ src/libertem/api.py
@@ -3,7 +3,7 @@ from typing import Union, Tuple
 
 import psutil
 import numpy as np
-from libertem.io.dataset import load, filetypes
+from libertem.io.dataset import load, filetypes, detect
 from libertem.io.dataset.base import DataSet
 from libertem.job.masks import ApplyMasksJob
 from libertem.job.raw import PickFrameJob
@@ -90,6 +90,37 @@ class Context:
         return load(filetype, executor=self.executor, *args, **kwargs)
 
     load.__doc__ = load.__doc__ % {"types": ", ".join(filetypes.keys())}
+
+    def dataset_info(self, path: str):
+        """
+        Try to detect the dataset at `path` and return some
+        diagnostics about it.
+
+        Note
+        ----
+        The information in the `diags` key is subject to change!
+
+        Parameters
+        ----------
+
+        path
+            Path to a file of the dataset you are interested in
+
+        Returns
+        -------
+        dict
+            Dictionary with keys `type`, `params` and `diags`.
+        """
+        params = detect(path, executor=self.executor)
+        ftype = params.pop('type', None)
+        if ftype is None:
+            return
+        ds = self.load(ftype, **params)
+        return {
+            'type': ftype,
+            'params': params,
+            'diags': ds.get_diagnostics(),
+        }
 
     def create_mask_job(self, factories, dataset, use_sparse=None,
                         mask_count=None, mask_dtype=None, dtype=None):
