### 2020-02-19 73488cb
Remove tileshape parameter from raw open form
