### 2020-02-19 e206513
Replace dataset_info with filetype="auto
For most file types, we don't need diagnostics, as we can just
open the DataSet and call `ds.get_diagnostics()`.
