+++ tests/io/test_base.py
@@ -3,6 +3,7 @@ from collections import namedtuple
 import numpy as np
 import pytest
 
+from libertem.io.dataset.base import DataSetException
 from libertem.io.dataset import get_extensions
 from libertem.io.dataset.base import (
     FileTree, Partition3D, _roi_to_nd_indices
@@ -130,3 +131,19 @@ def test_get_extensions():
     assert "mib" in exts
     assert "gtg" in exts
     # etc...
+
+
+def test_filetype_auto(hdf5, lt_ctx):
+    ds = lt_ctx.load("auto", path=hdf5.filename)
+    assert ds.ds_path == "data"
+
+def test_filetype_auto_fail_no_path(lt_ctx):
+    with pytest.raises(DataSetException) as e:
+        ds = lt_ctx.load("auto")
+    assert e.match("please specify the `path` kwarg to allow auto detection")
+
+
+def test_filetype_auto_fail_file_does_not_exist(lt_ctx):
+    with pytest.raises(DataSetException) as e:
+        ds = lt_ctx.load("auto", path="/does/not/exist/believe_me")
+    assert e.match("could not determine DataSet type for file")
