### 2020-02-19 f0e0a6e
Add documentation and changelog entry
