+++ docs/source/changelog/features/filetype_auto.rst
@@ -1,0 +1,5 @@
+[Feature] :code:`filetype="auto"` for Context.load
+==================================================
+
+* :meth:`libertem.api.Context.load` now automatically detects file
+  type and parameters if :code:`filetype="auto"` is passed (:pr:`621`)
