### 2020-02-19 8011473
Add documentation and changelog entry
