### 2020-02-19 7ae29c0
Tiled StdDevUDF
* Ensure that buffers are initialized on the control node, too
* Implement tiled processing with frame counter per tile region
