### 2020-02-19 0fcd708
Call preprocess on master node, too
Use case: initialization to enable simplifications of the merge
function, in case of a non-zero "neutral element."

Also clean up code duplication between run_for_dataset{,_async}
