### 2020-02-20 3a7a0ae
Tiled StdDevUDF
* Use process_tile()
* Per-tile frame counters
* Rely on new postprocess() behavior
* Reduce code duplication between StdDev and Clust
