### 2020-02-20 a6ef4f5
Tiled StdDevUDF
* Use process_tile()
* Per-tile frame counters
* Rely on new postprocess() behavior
* Reduce code duplication between StdDev and Clust
