### 2020-02-20 d0867ef
Changelog entry for StdDevUDF; documentation
