+++ docs/source/changelog/features/stddev.rst
@@ -1,0 +1,4 @@
+[Feature] :code:`StdDevUDF` speed-up
+====================================
+
+* More than 2x speed-up of standard deviation calculation with tiled processing (:pr:`625`)
