+++ src/libertem/io/dataset/memory.py
@@ -6,12 +6,39 @@ import numpy as np
 
 from libertem.web.messages import MessageConverter
 from libertem.io.dataset.base import (
-    FileSet3D, Partition3D, DataSet, DataSetMeta, DataTile
+    FileSet, BasePartition, DataSet, DataSetMeta, TilingScheme,
+    LocalFile, LocalFSMMapBackend,
 )
 from libertem.common import Shape
 
 
 log = logging.getLogger(__name__)
+
+
+class MemBackend(LocalFSMMapBackend):
+    def _set_readahead_hints(self, roi, fileset):
+        pass
+
+    def get_tiles(self, tiling_scheme, fileset, read_ranges, roi, native_dtype, read_dtype):
+        if roi is None:
+            # support arbitrary tiling in case of no roi
+            with fileset:
+                for tile in self._get_tiles_straight(tiling_scheme, fileset, read_ranges):
+                    yield tile.astype(read_dtype)
+        else:
+            # otherwise, require (1, ..., 1, N, Xi, ..., Xj) where the X match the ds shape
+            state = 0
+            for s1, s2 in zip(reversed(tiling_scheme.dataset_shape.sig),
+                              reversed(tiling_scheme.shape.sig)):
+                if state == 0:
+                    if s1 != s2:
+                        state = 1
+                        continue
+                elif state == 1:
+                    assert s2 == 1, "'non-contiguous' base_shape doesn't work with roi yet"
+            yield from super().get_tiles(
+                tiling_scheme, fileset, read_ranges, roi, native_dtype, read_dtype
+            )
 
 
 class MemDatasetParams(MessageConverter):
@@ -49,13 +76,11 @@ class MemDatasetParams(MessageConverter):
         return data
 
 
-class MemoryFile3D(object):
-    def __init__(self, data, check_cast=True):
-        self.num_frames = data.shape[0]
-        self.start_idx = 0
-        self.end_idx = self.num_frames
+class MemoryFile(LocalFile):
+    def __init__(self, data, check_cast=True, *args, **kwargs):
         self._data = data
         self._check_cast = check_cast
+        super().__init__(*args, **kwargs)
 
     def open(self):
         pass
@@ -63,14 +88,14 @@ class MemoryFile3D(object):
     def close(self):
         pass
 
-    def readinto(self, start, stop, out, crop_to=None):
-        slice_ = (...,)
-        if crop_to is not None:
-            slice_ = crop_to.get(sig_only=True)
-        if self._check_cast:
-            assert np.can_cast(self._data.dtype, out.dtype, casting='safe'),\
-                "cannot cast safely between %s and %s" % (self._data.dtype, out.dtype)
-        out[:] = self._data[(slice(start, stop),) + slice_]
+    def mmap(self):
+        return self._data
+
+    def raw_mmap(self):
+        return self._data.view(np.uint8)
+
+    def fileno(self):
+        raise NotImplementedError()
 
 
 class MemoryDataSet(DataSet):
@@ -88,22 +113,27 @@ class MemoryDataSet(DataSet):
     >>> ds = MemoryDataSet(data=data)
     '''
     def __init__(self, tileshape=None, num_partitions=None, data=None, sig_dims=2,
-                 check_cast=True, crop_frames=False, tiledelay=None, datashape=None):
+                 check_cast=True, tiledelay=None, datashape=None, base_shape=None,
+                 force_need_decode=False):
         # For HTTP API testing purposes: Allow to create empty dataset with given shape
         if data is None:
             assert datashape is not None
             data = np.zeros(datashape, dtype=np.float32)
         if num_partitions is None:
             num_partitions = psutil.cpu_count(logical=False)
-        if tileshape is None:
[…]
