### 2020-02-20 14c5144
Optimized tiled reading I/O infrastructure
 + Implement new I/O infrastructure
     + IOBackend, BufferPool, FileSet.get_read_ranges
 + Port DataSet implementations to the new I/O infrastructure
     + HDF5, SER: port, only local filesystem backend
     + SER: limited tiling support
     + MIB: port and add support for more raw formats (6bit, 12bit, 24bit)
     + K2IS: port, support for ROIs together with tiled reading
     + DM, FRMS6: port
     + memory: port, remove old crop_frames parameter,
 + Partition.get_macrotile: remove mmap parameter
 + Implement simple tileshape negotiation
     + 
[…]
