+++ src/libertem/udf/base.py
@@ -1,5 +1,6 @@
 from types import MappingProxyType
 from typing import Dict
+import logging
 import uuid
 
 import tqdm
@@ -9,6 +10,10 @@ import numpy as np
 from libertem.common.buffers import BufferWrapper, AuxBufferWrapper
 from libertem.common import Shape, Slice
 from libertem.utils.threading import set_num_threads
+from libertem.io.dataset.base import TilingScheme, Negotiator, Partition, DataSet
+
+
+log = logging.getLogger(__name__)
 
 
 class UDFMeta:
@@ -20,11 +25,14 @@ class UDFMeta:
         Added distinction of dataset_dtype and input_dtype
     """
     def __init__(self, partition_shape: Shape, dataset_shape: Shape, roi: np.ndarray,
-                 dataset_dtype: np.dtype, input_dtype: np.dtype):
+                 dataset_dtype: np.dtype, input_dtype: np.dtype, tiling_scheme: TilingScheme = None,
+                 tiling_index: int = 0):
         self._partition_shape = partition_shape
         self._dataset_shape = dataset_shape
         self._dataset_dtype = dataset_dtype
         self._input_dtype = input_dtype
+        self._tiling_scheme = tiling_scheme
+        self._tiling_index = tiling_index
         if roi is not None:
             roi = roi.reshape(dataset_shape.nav)
         self._roi = roi
@@ -56,6 +64,13 @@ class UDFMeta:
         Shape : The original shape of the whole dataset, not influenced by the ROI
         """
         return self._dataset_shape
+
+    @property
+    def tiling_scheme(self) -> Shape:
+        """
+        TilingScheme : the tiling scheme that was negotiated
+        """
+        return self._tiling_scheme
 
     @property
     def roi(self) -> np.ndarray:
@@ -409,6 +424,10 @@ class UDF(UDFBase):
     by overriding methods on this class.
     """
     USE_NATIVE_DTYPE = np.bool
+    TILE_SIZE_BEST_FIT = object()
+    TILE_SIZE_MAX = np.float("inf")
+    TILE_DEPTH_DEFAULT = object()
+    TILE_DEPTH_MAX = np.float("inf")
 
     def __init__(self, **kwargs):
         """
@@ -586,6 +605,12 @@ class UDF(UDFBase):
         '''
         return np.float32
 
+    def get_tiling_preferences(self):
+        return {
+            "depth": UDF.TILE_DEPTH_DEFAULT,
+            "total_size": UDF.TILE_SIZE_MAX,
+        }
+
     def cleanup(self):  # FIXME: name? implement cleanup as context manager somehow?
         pass
 
@@ -664,7 +689,7 @@ class Task(object):
 
 
 class UDFTask(Task):
-    def __init__(self, partition, idx, udf, roi):
+    def __init__(self, partition: Partition, idx, udf, roi):
         super().__init__(partition=partition, idx=idx)
         self._roi = roi
         self._udf = udf
@@ -681,7 +706,7 @@ class UDFRunner:
     def _get_dtype(self, dtype):
         return np.result_type(self._udf.get_preferred_input_dtype(), dtype)
 
-    def run_for_partition(self, partition, roi):
+    def run_for_partition(self, partition: Partition, roi):
         with set_num_threads(1):
             dtype = self._get_dtype(partition.dtype)
             meta = UDFMeta(
@@ -690,6 +715,7 @@ class UDFRunner:
                 roi=roi,
                 dataset_dtype=partition.dtype,
                 input_dtype=dtype,
+                tiling_scheme=None,
             )
             self._udf.set_meta(meta)
             self._udf.init_result_buffers()
@@ -699,21 +725,39 @@ class UDFRunner:
                 self._udf.clear_views()
[…]
