### 2020-02-20 8cd57e8
Actually remove tileshape from JSX
