### 2020-02-20 6c65765
Actually remove tileshape from JSX
