### 2020-02-20 215ef16
Work on Numba optimization of stddev
