### 2020-02-20 d665a89
Work on Numba optimization of stddev
