### 2020-02-21 05edd76
Improved docs for installation (#598)
* Improved docs for installation
* Added steps for installing LiberTEM for code contribution.
* Resources for beginners

Co-authored-by: Detective Loki <41474934+abiB27@users.noreply.github.com>
