+++ docs/source/changelog/documentation/install_doc.rst
@@ -1,0 +1,5 @@
+[Documentation] Added steps for installation
+============================================
+
+ * Better instructions for installing LiberTEM for
+   code contribution
