+++ docs/source/changelog/features/marker.rst
@@ -1,0 +1,4 @@
+[Feature] Improved marker
+======================================
+
+* Marker for point analysis and frame picking updated to allow precise aiming
