### 2020-02-22 23b6a3e
updated screenshot with new picker
