### 2020-02-22 e9afaee
Numba-based one-pass StdDevUDF
This version is slightly faster for the benign case of signal buffers
that fit comfortably into the CPU caches.

For the malign case of signal buffers that don't fit the caches well,
it is 5x-7x faster since it eliminates large intermediate buffers and
calculates each pixel in one pass.

* Added test for numerical stability.
* Added citation
