+++ tests/udf/test_stddev.py
@@ -1,23 +1,11 @@
 import pytest
 import numpy as np
+import numba
 
-from libertem.udf.stddev import run_stddev, tile_sum_var
+from libertem.udf.stddev import run_stddev, process_tile, merge
 from libertem.io.dataset.memory import MemoryDataSet
 
 from utils import _mk_random
-
-
-@pytest.mark.with_numba
-def test_tile_sum_var():
-    for i in range(100):
-        x = np.random.randint(1, 42)
-        y = np.random.randint(1, 42)
-        z = np.random.randint(1, 42)
-        data = _mk_random(size=(z, y, x), dtype="float32")
-        s, v = tile_sum_var(data)
-        assert np.allclose(s, np.sum(data, axis=0))
-        assert np.allclose(v, np.var(data, axis=0)*len(data))
-
 
 
 @pytest.mark.with_numba
@@ -55,6 +43,9 @@ def test_stddev(lt_ctx, use_roi):
     N = np.count_nonzero(roi)
     assert res['num_frame'] == N  # check the total number of frames
 
+    print(res['sum_frame'])
+    print(np.sum(data[roi], axis=0))
+    print(res['sum_frame'] - np.sum(data[roi], axis=0))
     assert np.allclose(res['sum_frame'], np.sum(data[roi], axis=0))  # check sum of frames
 
     assert np.allclose(res['mean'], np.mean(data[roi], axis=0))  # check mean
@@ -64,3 +55,60 @@ def test_stddev(lt_ctx, use_roi):
 
     std = np.std(data[roi], axis=0)
     assert np.allclose(std, res['std'])  # check standard deviation
+
+
+@numba.njit
+def _stability_workhorse(data):
+    s = np.zeros(1, dtype=np.float32)
+    varsum = np.zeros(1, dtype=np.float32)
+    N = 0
+    partitions = data.shape[0]
+    tiles = data.shape[1]
+    frames = data.shape[2]
+    for partition in range(partitions):
+        partition_sum = np.zeros(1, dtype=np.float32)
+        partition_varsum = np.zeros(1, dtype=np.float32)
+        partition_N = 0
+        for tile in range(tiles):
+            if partition_N == 0:
+                partition_sum[:] = np.sum(data[partition, tile])
+                partition_varsum[:] = np.var(data[partition, tile]) * frames
+                partition_N = frames
+            else:
+                partition_N = process_tile(
+                    tile=data[partition, tile],
+                    N0=partition_N,
+                    sum_inout=partition_sum,
+                    var_inout=partition_varsum
+                )
+        N = merge(
+            dest_N=N,
+            dest_sum=s,
+            dest_varsum=varsum,
+            src_N=partition_N,
+            src_sum=partition_sum,
+            src_varsum=partition_varsum
+        )
+    return N, s, varsum
+
+
+# This shouldn't be @pytest.mark.numba
+# since the calculation will take forever
+# with JIT disabled
+def test_stability(lt_ctx):
+    """
+    Test variance, standard deviation, sum of frames, and mean computation
+    implemented in udf/stddev.py
+
+    Parameters
+    ----------
+    lt_ctx
+        Context class for loading dataset and creating jobs on them
+    """
+    data = _mk_random(size=(1024, 1024, 8, 1), dtype="float32")
+
+    N, s, varsum = _stability_workhorse(data)
+
+    assert N == np.prod(data.shape)
+    assert np.allclose(data.sum(), s)
+    assert np.allclose(data.var(ddof=N-1), varsum)
