+++ src/libertem/udf/stddev.py
@@ -1,118 +1,115 @@
-import collections
-
 import numpy as np
 import numba
 
 from libertem.udf import UDF
 
 
-VariancePart = collections.namedtuple('VariancePart', ['var', 'sum_im', 'N'])
+@numba.njit(fastmath=True)
+def _merge(N, N0, sum_im0, var0, N1, sum_im1, var1):
+    '''
+    Basic function to perform numerically stable merge :cite:`Schubert2018`
+    '''
+    # compute mean for each partitions
+    mean_A = sum_im0 / N0
+    mean_B = sum_im1 / N1
+
+    # compute mean for joint samples
+    delta = mean_B - mean_A
+    mean = mean_A + (N1 * delta) / N
+
+    # compute sum of images for joint samples
+    sum_im_AB = sum_im0 + sum_im1
+
+    # compute sum of variances for joint samples
+    delta_P = mean_B - mean
+    var_AB = var0 + var1 + (N1 * delta * delta_P)
+    return sum_im_AB, var_AB
 
 
-@numba.njit
-def merge(N0, sum_im0, var0, N1, sum_im1, var1):
+@numba.njit(fastmath=True)
+def merge(dest_N, dest_sum, dest_varsum, src_N, src_sum, src_varsum):
     """
     Given two sets of partitions, with sum of frames
-    and sum of variances, compute joint sum of frames
-    and sum of variances using one pass algorithm
+    and sum of variances, aggregate joint sum of frames
+    and sum of variances in destination partition using one pass
+    algorithm :cite:`Schubert2018`
 
     Parameters
     ----------
-    N0, sum_im0, var0
-        Contains information about the first partition:
-        Number of frames used, sum of pixels, sum of variances
+    dest_N : int
+        Number of frames aggregated in dest_sum and dest_varsum
+        The case :code:`dest_N == 0` is handled correctly.
+    dest_sum, dest_varsum
+        Aggregation buffers, will be updated: sum of pixels, sum of variances
+    src_N, src_sum, src_varsum
+        Information about the second partition
 
-    N1, sum_im1, var1
-        Contains information about the second partition:
-        
 
     Returns
     -------
-    N, sum_im, var:
-        Number of frames used, sum of pixels, sum of variances of the merged partitions
+    N:
+        New number of frames aggregated in aggregation buffer
     """
-    if N0 == 0:
-        return N1, sum_im1, var1
-    if N1 == 0:
-        return N0, sum_im0, var0
-    N = N0 + N1
+    if dest_N == 0:
+        dest_sum[:] = src_sum
+        dest_varsum[:] = src_varsum
+        return src_N
+    else:
+        N = dest_N + src_N
 
-    shape = sum_im0.shape
-
-    sum_im0_ = sum_im0.flatten()
-    sum_im1_ = sum_im1.flatten()
-    sum_im_AB = np.zeros_like(sum_im0_)
-    var0_ = var0.flatten()
-    var1_ = var1.flatten()
-    var_AB = np.zeros_like(var0_)
-
-    le = len(sum_im0_)
-    BLOCKSIZE = 16
-    n_blocks = le // BLOCKSIZE
-
-    def _merge(sum_im0, var0, sum_im1, var1):
-        # compute mean for each partitions
-        mean_A = sum_im0 / N0
-        mean_B = sum_im1 / N1
-
-        # compute mean for joint samples
-        delta = mean_B - mean_A
-        mean = mean_A + (N1 * delta) / N
-
-        # compute sum of images for joint samples
-        sum_im_AB = sum_im0 + sum_im1
-
-        # compute sum of variances for joint samples
-        delta_P = mean_B - mean
-        var_AB = var0 + var1 + (N1 * delta * delta_P)
[…]
