+++ docs/source/references-libertem.bib
@@ -1117,4 +1117,14 @@
 	pages = {016102},
 	file = {Electron holography—basics and applications - IOPscience:/Users/migunov/Zotero/storage/7C3UUE82/meta\;jsessionid=3F5DA019F9A86C312BFD8889CFC17030.c2.iopscience.cld.iop.html:text/html;Lichte_H,Lehmann_M-E-holog-basics&app(2008).pdf:/Users/migunov/Zotero/storage/WRB52ACC/Lichte_H,Lehmann_M-E-holog-basics&app(2008).pdf:application/pdf}
 }
+
+@InProceedings{Schubert2018,
+  author    = {Erich Schubert and Michael Gertz},
+  title     = {Numerically stable parallel computation of (co-)variance},
+  booktitle = {Proceedings of the 30th International Conference on Scientific and Statistical Database Management - {SSDBM} {\textquotesingle}18},
+  year      = {2018},
+  publisher = {{ACM} Press},
+  doi       = {10.1145/3221269.3223036},
+}
+
 @Comment{jabref-meta: databaseType:bibtex;}
