+++ docs/source/changelog/features/marker.rst
@@ -1,4 +1,4 @@
-[Feature] Improved marker
+[Feature] Improved pick marker
 ======================================
 
-* Marker for point analysis and frame picking updated to allow precise aiming
+* Marker for point analysis and frame picking updated to allow precise aiming (:pr:`639`)
