### 2020-02-22 34e3604
Fix issue 600
Better error message when parameter detector_size
is missing.
