### 2020-02-23 1197c31
Added two tests
1. test to check error message when missing
parameter detector_size
2. test to check, opening raw file old way with
detector_size_raw and crop_detector_to
