+++ src/libertem/io/dataset/raw.py
@@ -191,8 +191,8 @@ class RawFileDataSet(DataSet):
             warnings.warn("tileshape argument is deprecated, ignored", DeprecationWarning)
 
         if crop_detector_to is not None:
-            warnings.warn("crop_detector_to and detector_size_raw are deprecated, ",
-                          "and will be removed after version 0.6.0. ",
+            warnings.warn("crop_detector_to and detector_size_raw are deprecated, "
+                          "and will be removed after version 0.6.0. "
                           "please specify detector_size instead or use EMPAD DataSet",
                           DeprecationWarning)
             if detector_size is not None:
