+++ docs/source/changelog/features/initial_dataset_name.rst
@@ -1,0 +1,4 @@
+[Feature] Initially set file's basename as dataset name in GUI
+======================================
+
+* When opening a file with the GUI for the first time, dataset name is initially set to the file's basename (:pr:`642`)
