### 2020-02-23 df1623b
Added types to libertem.api.Context
