+++ src/libertem/api.py
@@ -9,15 +9,16 @@ from libertem.job.masks import ApplyMasksJob
 from libertem.job.raw import PickFrameJob
 from libertem.job.base import Job
 from libertem.common import Slice, Shape
+from libertem.common.buffers import BufferWrapper
 from libertem.executor.dask import DaskJobExecutor
-from libertem.analysis.raw import PickFrameAnalysis
-from libertem.analysis.com import COMAnalysis
-from libertem.analysis.radialfourier import RadialFourierAnalysis
+from libertem.analysis.raw import PickFrameAnalysis, PickResultSet
+from libertem.analysis.com import COMAnalysis, COMResultSet
+from libertem.analysis.radialfourier import RadialFourierAnalysis, RadialFourierResultSet
 from libertem.analysis.disk import DiskMaskAnalysis
 from libertem.analysis.ring import RingMaskAnalysis
-from libertem.analysis.sum import SumAnalysis
+from libertem.analysis.sum import SumAnalysis, SumResultSet
 from libertem.analysis.point import PointMaskAnalysis
-from libertem.analysis.masks import MasksAnalysis
+from libertem.analysis.masks import MasksAnalysis, MasksResultSet, SingleMaskResultSet
 from libertem.analysis.base import Analysis
 from libertem.udf.base import UDFRunner, UDF
 from libertem.udf.auto import AutoUDF
@@ -95,7 +96,7 @@ class Context:
     load.__doc__ = load.__doc__ % {"types": ", ".join(filetypes.keys())}
 
     def create_mask_job(self, factories, dataset, use_sparse=None,
-                        mask_count=None, mask_dtype=None, dtype=None):
+                        mask_count=None, mask_dtype=None, dtype=None) -> np.ndarray:
         """
         Create a low-level mask application job. Each factory function should, when called,
         return a numpy array with the same shape as frames in the dataset (so dataset.shape.sig).
@@ -170,7 +171,7 @@ class Context:
         )
 
     def create_mask_analysis(self, factories, dataset, use_sparse=None,
-                             mask_count=None, mask_dtype=None, dtype=None):
+                             mask_count=None, mask_dtype=None, dtype=None) -> MasksResultSet:
         """
         Create a mask application analysis. Each factory function should, when
         called, return a numpy array with the same shape as frames in the
@@ -246,7 +247,7 @@ class Context:
                 "dtype": dtype},
         )
 
-    def create_com_analysis(self, dataset, cx: int = None, cy: int = None, mask_radius: int = None):
+    def create_com_analysis(self, dataset, cx: int = None, cy: int = None, mask_radius: int = None) -> COMResultSet:
         """
         Create a center-of-mass (first moment) analysis, possibly masked.
 
@@ -282,7 +283,7 @@ class Context:
 
     def create_radial_fourier_analysis(self, dataset, cx: float = None, cy: float = None,
             ri: float = None, ro: float = None, n_bins: int = None, max_order: int = None,
-            use_sparse: bool = None):
+            use_sparse: bool = None) -> RadialFourierResultSet:
         """
         Create an Analysis that calculates the Fourier transform of rings around the center.
 
@@ -324,7 +325,7 @@ class Context:
         )
         return analysis
 
-    def create_disk_analysis(self, dataset, cx: int = None, cy: int = None, r: int = None):
+    def create_disk_analysis(self, dataset, cx: int = None, cy: int = None, r: int = None) -> SingleMaskResultSet:
         """
         Create an Analysis that integrates over a disk (i.e. filled circle).
 
@@ -354,7 +355,7 @@ class Context:
         )
 
     def create_ring_analysis(
[…]
