### 2020-02-23 c123654
Made a function to fix issues
