### 2020-02-23 6f36d66
Updated api.py with exact return type
