### 2020-02-23 1751c16
Improved fix 600
Added tests to test_raw.py
Changes ValueError to TypeError
