### 2020-02-23 9332832
Improved fix 600
Added tests to test_raw.py
Changes ValueError to TypeError
