### 2020-02-23 2a72683
Updated return type of run(), run_udf(), create_mask_job() and added name in creators.json
