### 2020-02-23 a28117d
Updated Dict type and creators.json to original
