### 2020-02-24 9031cd1
added parameter type to libertem.api.Context
