+++ src/libertem/api.py
@@ -1,5 +1,5 @@
 import warnings
-from typing import Union, Tuple, Dict
+from typing import Union, Tuple, Dict, Iterable, Callable
 
 import psutil
 import numpy as np
@@ -11,6 +11,7 @@ from libertem.job.base import Job
 from libertem.common import Slice, Shape
 from libertem.common.buffers import BufferWrapper
 from libertem.executor.dask import DaskJobExecutor
+from libertem.executor.base import JobExecutor
 from libertem.analysis.raw import PickFrameAnalysis
 from libertem.analysis.com import COMAnalysis
 from libertem.analysis.radialfourier import RadialFourierAnalysis
@@ -31,7 +32,7 @@ class Context:
     them.
     """
 
-    def __init__(self, executor=None):
+    def __init__(self, executor: JobExecutor=None):
         """
         Create a new context. In the background, this creates a suitable
         executor and spins up a local Dask cluster.
@@ -96,7 +97,7 @@ class Context:
     load.__doc__ = load.__doc__ % {"types": ", ".join(filetypes.keys())}
 
     def create_mask_job(self, factories, dataset, use_sparse=None,
-                        mask_count=None, mask_dtype=None, dtype=None) -> ApplyMasksJob:
+                        mask_count=None, mask_dtype=None,dtype=None) -> ApplyMasksJob:
         """
         Create a low-level mask application job. Each factory function should, when called,
         return a numpy array with the same shape as frames in the dataset (so dataset.shape.sig).
@@ -170,8 +171,9 @@ class Context:
             dtype=dtype,
         )
 
-    def create_mask_analysis(self, factories, dataset, use_sparse=None,
-                             mask_count=None, mask_dtype=None, dtype=None) -> MasksAnalysis:
+    def create_mask_analysis(self, factories, dataset: DataSet, use_sparse: bool = None,
+            mask_count: int = None, mask_dtype: np.dtype = None,
+            dtype: np.dtype = None) -> MasksAnalysis:
         """
         Create a mask application analysis. Each factory function should, when
         called, return a numpy array with the same shape as frames in the
@@ -247,7 +249,7 @@ class Context:
                 "dtype": dtype},
         )
 
-    def create_com_analysis(self, dataset, cx: int = None, cy: int = None,
+    def create_com_analysis(self, dataset: DataSet, cx: int = None, cy: int = None,
                             mask_radius: int = None) -> COMAnalysis:
         """
         Create a center-of-mass (first moment) analysis, possibly masked.
@@ -282,7 +284,7 @@ class Context:
         )
         return analysis
 
-    def create_radial_fourier_analysis(self, dataset, cx: float = None, cy: float = None,
+    def create_radial_fourier_analysis(self, dataset: DataSet, cx: float = None, cy: float = None,
             ri: float = None, ro: float = None, n_bins: int = None, max_order: int = None,
             use_sparse: bool = None) -> RadialFourierAnalysis:
         """
@@ -326,7 +328,7 @@ class Context:
         )
         return analysis
 
-    def create_disk_analysis(self, dataset, cx: int = None, cy: int = None,
+    def create_disk_analysis(self, dataset: DataSet, cx: int = None, cy: int = None,
                              r: int = None) -> DiskMaskAnalysis:
         """
         Create an Analysis that integrates over a disk (i.e. filled circle).
@@ -356,7 +358,7 @@ class Context:
             dataset=dataset, parameters=parameters
         )
 
-    def create_ring_analysis(self, dataset, cx: int = None, cy: int = None,
+    def create_ring_analysis(self, dataset: DataSet, cx: int = None, cy: int = None,
[…]
