### 2020-02-24 6f1c865
Updated parameter types and rebase to master
