### 2020-02-24 275f486
Try citing in main text
A second pass or citation in main text seem to fix the issue with undefined reference -- perhaps
BibTeX cache not generated when citing in docstring?
