### 2020-02-24 a177aea
Remove LaTeX from BibTeX
