+++ docs/source/references-libertem.bib
@@ -1121,7 +1121,7 @@
 @InProceedings{Schubert2018,
   author    = {Erich Schubert and Michael Gertz},
   title     = {Numerically stable parallel computation of (co-)variance},
-  booktitle = {Proceedings of the 30th International Conference on Scientific and Statistical Database Management - {SSDBM} {\textquotesingle}18},
+  booktitle = {Proceedings of the 30th International Conference on Scientific and Statistical Database Management - {SSDBM} 18},
   year      = {2018},
   publisher = {{ACM} Press},
   doi       = {10.1145/3221269.3223036},
