### 2020-02-24 7ffbeb5
Try not citing in docstring
CI error not reproducible locally.
