+++ docs/source/reference/udf.rst
@@ -62,7 +62,7 @@ Some generally useful UDFs are included with LiberTEM:
 
 .. _`stddev udf`:
 
-Single-pass standard deviation based on :cite:`Schubert2018`.
+Single-pass standard deviation based on Schubert2018.
 
 .. autoclass:: libertem.udf.stddev.StdDevUDF
     :members:
