### 2020-02-24 f3861aa
Manual citation due to issues with sphinxcontrib-bibtex
:cite:`Schubert2018` glitches in CI in docstrings and some RST files.

Hard to reproduce the CI behavior
