### 2020-02-24 356079b
Updated with MaskArrayType
