### 2020-02-25 a45c51c
Update mask.py and api.py with right type alias
