### 2020-02-25 fc1bf80
Update mask.py and api.py with right type alias
