+++ src/libertem/io/dataset/raw.py
@@ -6,6 +6,7 @@ import numpy as np
 
 from libertem.common import Shape
 from libertem.web.messages import MessageConverter
+from libertem.common.buffers import zeros_aligned
 from .base import (
     DataSet, DataSetException, DataSetMeta,
     Partition3D, File3D, FileSet3D,
@@ -286,7 +287,8 @@ class RawFileDataSet(DataSet):
         for part_slice, start, stop in Partition3D.make_slices(
                 shape=self.shape,
                 num_partitions=self._get_num_partitions()):
-            yield Partition3D(
+            yield RawPartition(
+                enable_direct=self._enable_direct,
                 stackheight=self._stackheight,
                 meta=self._meta,
                 fileset=fileset,
@@ -297,3 +299,14 @@ class RawFileDataSet(DataSet):
 
     def __repr__(self):
         return "<RawFileDataSet of %s shape=%s>" % (self.dtype, self.shape)
+
+
+class RawPartition(Partition3D):
+    def __init__(self, enable_direct, *args, **kwargs):
+        self._enable_direct = enable_direct
+        super().__init__(*args, **kwargs)
+
+    def zeros(self, *args, **kwargs):
+        if self._enable_direct:
+            return zeros_aligned(*args, **kwargs)
+        return super().zeros(*args, **kwargs)
