### 2020-02-25 12a2b9b
Fix performance issue w/ large tiles
Allocating large buffers with zeros_aligned is slower
than using np.zeros, but is still needed for direct I/O.
This change only enables zeros_aligned for direct I/O and also
implements a shortcut in _get_tiles_with_roi if the roi is all-zeros
