### 2020-02-25 d34755b
skip direct I/O test on windows
