### 2020-02-25 523b33b
Added citation of dataset used
Citation for dataset used in screenshot
is added.
