### 2020-02-26 77e7dcd
Add items to release checklist
