### 2020-02-26 faa3794
Add items to release checklist
Added items mentioned in issue #629
