### 2020-02-26 fa00d63
Added proper citation
Removed previous changes, added
citation in a systematic way
