### 2020-02-26 116fe97
Added proper citation
Removed previous changes, added
citation in a systematic way
