+++ docs/source/references-libertem.bib
@@ -1128,3 +1128,15 @@
 }
 
 @Comment{jabref-meta: databaseType:bibtex;}
+
+@dataset{ophus_colin_2019_3592520,
+  author       = {Ophus, Colin and
+                  Savitzky, Benjamin},
+  title        = {{Simulated calibration dataset for 4D scanning 
+                   transmission electron microscopy}},
+  month        = dec,
+  year         = 2019,
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.3592520},
+  url          = {https://doi.org/10.5281/zenodo.3592520}
+}
