### 2020-02-26 d0d2b49
Initially set file's basename as dataset name (#642)
* Fixed #246
* Added to changelog
* Fixed raw file issue, rebuilt client and updated creators.json
