+++ client/src/dataset/helpers.ts
@@ -1,4 +1,5 @@
 import { withFormik } from "formik";
+import * as pathfind from 'path';
 import { DatasetTypes } from "../messages";
 import { OpenFormProps } from "./types";
 import { validateOpen } from "./validate";
@@ -12,6 +13,14 @@ export function getInitial<T extends object, K extends keyof T, V>(key: K, other
         return otherwise;
     }
     const res = values[key] !== undefined ? values[key] : otherwise;
+    return res;
+}
+
+export function getInitialName<T extends object, K extends keyof T>(key: K, otherwise: string, values?: T): string | T[K] {
+    if (!values) {
+        return pathfind.basename(otherwise);
+    }
+    const res = values[key] !== undefined ? values[key] : pathfind.basename(otherwise);
     return res;
 }
 
