### 2020-02-26 7b50762
Add items in right order
Placed items at proper order under 'Before'.
