### 2020-02-28 1a5f116
Fix DOI on citing page
Use permanent DOI instead of 0.1.0
