+++ docs/source/citing.rst
@@ -4,7 +4,7 @@ Citing
 To help us spread the word, please credit and cite LiberTEM in publications where it has been significant. 
 Resources for citing LiberTEM are linked through the DOI badge below.
 
-.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.1478763.svg
-   :target: https://doi.org/10.5281/zenodo.1478763
+.. |zenodo| image:: https://zenodo.org/badge/DOI/10.5281/zenodo.1477847.svg
+.. _zenodo: https://doi.org/10.5281/zenodo.1477847
 
 Please refer to our :ref:`authorship` for details regarding authorship and credit for contributors.
