### 2020-02-28 f51ab00
Review @sk1p; further speed-up
Thx @sk1p for the feedback!
