+++ src/libertem/udf/stddev.py
@@ -5,45 +5,14 @@ from libertem.udf import UDF
 
 
 @numba.njit(fastmath=True)
-def _merge(N, N0, sum_im0, var0, N1, sum_im1, var1):
+def merge_single(n, n_0, sum_0, varsum_0, n_1, sum_1, varsum_1, mean_1):
     # FIXME manual citation due to issues in CI.
     # Check if :cite:`Schubert2018` works in a future release
     '''
-    Basic function to perform numerically stable merge
+    Basic function to perform numerically stable merge.
 
-    Erich Schubert and Michael Gertz. Numerically stable parallel computation of
-    (co-)variance. In `Proceedings of the 30th International Conference on
-    Scientific and Statistical Database Management - SSDBM 18`. ACM Press, 2018.
-    `doi:10.1145/3221269.3223036 <https://doi.org/10.1145/3221269.3223036>`_
-
-    cite:Schubert2018
-    '''
-    # compute mean for each partitions
-    mean_A = sum_im0 / N0
-    mean_B = sum_im1 / N1
-
-    # compute mean for joint samples
-    delta = mean_B - mean_A
-    mean = mean_A + (N1 * delta) / N
-
-    # compute sum of images for joint samples
-    sum_im_AB = sum_im0 + sum_im1
-
-    # compute sum of variances for joint samples
-    delta_P = mean_B - mean
-    var_AB = var0 + var1 + (N1 * delta * delta_P)
-    return sum_im_AB, var_AB
-
-
-@numba.njit(fastmath=True)
-def merge(dest_N, dest_sum, dest_varsum, src_N, src_sum, src_varsum):
-    # FIXME manual citation due to issues in CI.
-    # Check if :cite:`Schubert2018` works in a future release
-    """
-    Given two sets of partitions, with sum of frames
-    and sum of variances, aggregate joint sum of frames
-    and sum of variances in destination partition using one pass
-    algorithm.
+    This function is designed to be inlined in a loop over all pixels in a frame
+    to merge an individual pixel, with some parts pre-calculated.
 
     Erich Schubert and Michael Gertz. Numerically stable parallel computation of
     (co-)variance. In `Proceedings of the 30th International Conference on
@@ -54,13 +23,72 @@ def merge(dest_N, dest_sum, dest_varsum, src_N, src_sum, src_varsum):
 
     Parameters
     ----------
-    dest_N : int
+    n : int
+        Total number of frames, assumed n == n_0 + n_1
+        Pre-calculated since equal for all pixels.
+    n_0 : int
+        Number of frames aggregated in sum_0 and varsum_0, n_0 > 0.
+        The case n_0 == 0 is handled separately in the calling function.
+    sum_0, varsum_0 : float
+        Aggregate sum and sum of variances
+    n_1 : int
+        Number of frames to merge from minibatch, n_1 > 0
+    sum_1, varsum_1, mean_1 : float
+        Minibatch sum, sum of variances and mean. The mean is
+        available from the previous minibatch calculation in the innermost
+        aggregation loop as a "hot" value.
+
+    Returns
+    -------
+    sum, varsum : float
+        New aggregate sum and sum of variances
+    '''
+    # compute mean for each partitions
+    mean_0 = sum_0 / n_0
+
+    # compute mean for joint samples
+    delta = mean_1 - mean_0
+    mean = mean_0 + (n_1 * delta) / n
+
+    # compute sum of images for joint samples
+    sumsum = sum_0 + sum_1
+
+    # compute sum of variances for joint samples
+    partial_delta = mean_1 - mean
+    varsum = varsum_0 + varsum_1 + (n_1 * delta * partial_delta)
[…]
