+++ docs/source/changelog/misc/stddev.rst
@@ -1,0 +1,5 @@
+[Misc] Rename result buffers of StdDevUDF
+=========================================
+
+* Rename result buffers of :class:`~libertem.udf.stddev.StdDevUDF`
+  from :code:`'sum_frame'` to :code:`'sum'`, :code:`'num_frame'` to :code:`'num_frames'` (:pr:`640`)
