### 2020-02-28 aa682ef
Review @sk1p; further speed-up
Thx @sk1p for the feedback!
