+++ src/libertem/analysis/sd.py
@@ -14,6 +14,7 @@ class SDAnalysis(BaseAnalysis):
         return get_roi(params=self.parameters, shape=self.dataset.shape.nav)
 
     def get_udf_results(self, udf_results, roi):
+        udf_results = std.consolidate_result(udf_results)
         return AnalysisResultSet([
             AnalysisResult(raw_data=udf_results['var'].data,
                            visualized=visualize_simple(
