### 2020-02-28 cd99a2e
Clarify changelog entry, include var -> varsum
