### 2020-02-28 6673aa3
Ensure coverage of blocked loop
The small previous shape only covered the tail handling.
