### 2020-02-29 6a928c7
Added progress bar support
Added progress bar support for `Context.map` and `Context.run`
