+++ docs/source/changelog/misc/progress_bar_support.rst
@@ -1,0 +1,5 @@
+[Misc] Added progress bar support
+=================================
+
+* Added progress bar support for :meth:`libertem.api.Context.run` and
+  :meth:`libertem.api.Context.map`
