### 2020-02-29 4bc9b22
Added progress bar support
Added progress bar support for `Context.map` and `Context.run`
