### 2020-02-29 ef536dd
Made progress bar optional
Added parameter `progress`, default False.
