### 2020-03-01 6b6efe9
Exception for Direct I/O for raw files on Windows
