+++ docs/source/changelog/bugfix/raw_direct_io_windows.rst
@@ -1,0 +1,4 @@
+[Feature] Exception for Direct I/O for raw files on Windows
+============================================================
+
+* Exception is raised when Direct I/O is enabled for raw files on Windows (:pr:`659`)
