### 2020-03-01 LiberTEM/LiberTEM#660 @fedorov
Short installation instructions do not work
Re: JOSS review in https://github.com/openjournals/joss-reviews/issues/2006

Following the instructions in README, I have the following issues:
* invalid python binary (I do have python 3.7.6, and I was able to run the command using pyton3 instead)
```bash
$ virtualenv -p python3.6 ~/libertem-venv/
The path python3.6 (from --python=python3.6) does not exist
(libertem-venv)
```
* invalid pip command; I was able to run `pip install librtem`
```bash
$ pip install libertem[torch]
zsh: no matches found: libertem[torch]
(libertem-venv)
```
