### 2020-03-01 #2006 @fedorov
@uellue very nice work! I submitted just one issue, and also have few clarification questions below, before I check off the remaining items in the checklist (am not sure if I should submit issues about those):
* I could not find (or missed) pointers to publicly available sample datasets that software users could easily access - would you be able to add some pointers to the documentation?
* there are statements claiming functionality and performance advantages over alternatives, but it is not clear if those were confirmed anywhere. Do you have any additional citations comparing performance with other packages? I am aware of the [performance section of the documentation](https://libertem.github.io/LiberTEM/performance.html), but the paper mentions few other packages, and the performance section does not include comparison with those. Since there are statements claiming superiority in the paper, it would be good to back them with evidence.
* I think it would be helpful to the user if the documentation contained a section listing similar packages (something along the lines that you already have in the submitted paper PDF). Would it make sense to add that section to the documentation?
