### 2020-03-01 LiberTEM/LiberTEM#660 @sk1p
Thanks for the report! Yes, we may need to clarify that a supported Python version needs to be substituted for `python3.6` in that command.

> `zsh: no matches found: libertem[torch]`

That appears to be a quirk of zsh, should be fixable by quoting the package name argument.
