### 2020-03-02 LiberTEM/LiberTEM#660 @fedorov
I am sure LiberTEM users are sophisticated enough that they will figure this out without problem, so this is more for "looks" to have the front-page instructions more clean.
