+++ src/libertem/web/state.py
@@ -1,0 +1,197 @@
+import os
+import typing
+
+import psutil
+
+import libertem
+from libertem.analysis.base import AnalysisResultSet
+from libertem.io.dataset.base import DataSetException
+
+
+class ExecutorState:
+    def __init__(self):
+        self.executor = None
+
+    def get_executor(self):
+        if self.executor is None:
+            # TODO: exception type, conversion into 400 response
+            raise RuntimeError("wrong state: executor is None")
+        return self.executor
+
+    def have_executor(self):
+        return self.executor is not None
+
+    async def set_executor(self, executor, params):
+        if self.executor is not None:
+            await self.executor.close()
+        self.executor = executor
+        self.cluster_params = params
+
+
+AnalysisParameters = typing.Dict
+
+
+class AnalysisState:
+    def __init__(self, executor_state: ExecutorState):
+        self.analyses: typing.Dict[str, AnalysisParameters] = {}
+        self.results: typing.Dict[str, typing.Tuple[AnalysisResultSet, AnalysisParameters]] = {}
+
+    def create_analysis(self, uuid, dataset_uuid, analysis_type, parameters):
+        assert uuid not in self.analyses
+        self.analyses[uuid] = {
+            "type": analysis_type,
+            "dataset": dataset_uuid,
+            "params": parameters,
+        }
+
+    def get_analysis(self, uuid):
+        return self.analyses.get(uuid)
+
+
+class DatasetState:
+    def __init__(self, executor_state: ExecutorState):
+        self.datasets = {}
+        self.executor_state = executor_state
+
+    def register(self, uuid, dataset, params):
+        assert uuid not in self.datasets
+        self.datasets[uuid] = {
+            "dataset": dataset,
+            "params": params,
+        }
+        self.dataset_to_id[dataset] = uuid
+        return self
+
+    async def serialize(self, dataset_id):
+        executor = self.executor_state.get_executor()
+        dataset = self.datasets[dataset_id]
+        diag = await executor.run_function(lambda: dataset["dataset"].diagnostics)
+        return {
+            "id": dataset_id,
+            "params": {
+                **dataset["params"]["params"],
+                "shape": tuple(dataset["dataset"].shape),
+            },
+            "diagnostics": diag,
+        }
+
+    async def serialize_all(self):
+        executor = self.executor_state.get_executor()
+        return [
+            await self.serialize_dataset(executor, dataset_id)
+            for dataset_id in self.datasets.keys()
+        ]
+
+    def __getitem__(self, uuid):
+        return self.datasets[uuid]["dataset"]
+
+    def __contains__(self, uuid):
+        return uuid in self.datasets
+
+    async def verify(self):
+        executor = self.executor_state.get_executor()
+        for uuid, params in self.datasets.items():
+            dataset = params["dataset"]
+            try:
+                await executor.run_function(dataset.check_valid)
+            except DataSetException:
+                await self.remove_dataset(uuid)
+
+    async def remove(self, uuid):
+        ds = self.datasets[uuid]["dataset"]
+        jobs_to_remove = [
+            job
+            for job in self.jobs.values()
+            if self.dataset_to_id[self.dataset_for_job[job]] == uuid
+        ]
+        job_ids = {self.job_to_id[job]
+                   for job in jobs_to_remove}
+        del self.datasets[uuid]
+        del self.dataset_to_id[ds]
+        for job_id in job_ids:
+            await self.remove_job(job_id)
+
+
+class JobState:
+    def __init__(self, executor_state: ExecutorState, dataset_state: DatasetState):
+        self.jobs = {}
+        self.job_to_id = {}
+        self.executor_state = executor_state
+        self.dataset_state = dataset_state
+
[…]
