+++ tox.ini
@@ -132,8 +132,6 @@ deps=
     nbsphinx
     nbsphinx_link
     ipython
-    # Inject release candidates for doctest testing
-    -roverride_requirements.txt
 skipsdist=True
 whitelist_externals=
     cat

[…]
