### 2020-03-02 841febc
Comment out index URLs from override_requirements
Needs to stay this way as long as the rest of the file is empty,
otherwise all dependencies are installed via those indexes under
some circumstances.
