### 2020-03-02 LiberTEM/LiberTEM#554 @uellue
The new `StdDevUDF` #640 could be a good test bed for the impact of tile shape negotiation. It passes twice over a subset of a tile for the "minibatch" two-pass calculation of sum and sum of variances. If a memory-mapped tile contains so many frames that `n_frames * BLOCKSIZE` doesn't fit the L3 cache, it will likely lose performance dramatically. It could even make sense to restrict the number of frames so that `n_frames * BLOCKSIZE` fits the L2 cache.

Being able to give a size hint for the navigation size of a tile saves the additional complication of implementing blocking in navigation dimension. Instead, this level is then delegated to the LiberTEM back-end. The requirements for the `StdDevUDF` would be to specify a minimum and maximum comfortable nav size for a tile, while the optimal signal size can be decided by the back end.

If the data is converted, the total tile size should fit the L3 cache. If the data is memory-mapped, tiles can have arbitrarily large signal size.
