### 2020-03-04 LiberTEM/LiberTEM#660 @sk1p
> why we need to quote the package name ?

The error happens because the zsh shell used here interprets the brackets as a pattern (either glob or regular expression, I'm not sure as I use a different shell) and tries to find a file that matches the  pattern. Adding quotes removes this meaning from the brackets and passes the whole string down literally to `pip`.
