### 2020-03-05 LiberTEM/LiberTEM#660 @sk1p
Fixed by #664
