### 2020-03-09 LiberTEM/LiberTEM#554 @uellue
Discussion with @sk1p: We should check if using native input dtype in StdDevUDF and casting only within `process_tile()` brings a performance benefit once tile shape negotiation works. The values are up-casted to `float64`anyway now for stability reasons #672.
