### 2020-03-20 LiberTEM/LiberTEM#554 @uellue
`scipy.sparse.csr_matrix` doesn't work well at all on large memory-mapped partitions, likely for similar reasons as StdDevUDF. This could be another test case and use case, in particular since the `ApplyMasksUDF` will have to signal a tile shape depending on the current back-end:

* OpenBLAS or Torch for dense masks: No restrictions.
* scipy.sparse: Cache-friendly tile height, might also depend on number of masks
