### 2020-03-20 #2006 @uellue
Hi @fedorov @majensen FYI, I have been working on a deeper comparison between Dask and LiberTEM both in programming model and benchmarking. That's still WIP. Two things can already be said: First, Dask has improved impressively since I last looked, and second, the comparison has already helped us to identify opportunities for improvement: https://github.com/LiberTEM/LiberTEM/issues/554#issuecomment-601668112

In that sense, many thanks! More details to follow when the comparison is done. :-)
