### 2020-04-06 LiberTEM/LiberTEM#554 @uellue
Discussion with @sk1p:

* We can leave the signal size open as a free variable since this is the "fast" axis and typical optimized algorithms do their own blocking for L1 cache friendliness in this direction.
* The total tile size in bytes only has a maximum, no minimum: Always as large as possible to reduce overhead in LiberTEM!

In conclusion, remaining are nav size min and max, and total size max.
