### 2020-04-06 LiberTEM/LiberTEM#554 @uellue
Discussion with @sk1p: We need diagnostic tools.

* Negotiation inputs and results
* I/O method
* Decoding or not
