### 2020-04-06 LiberTEM/LiberTEM#554 @uellue
Information required for size negotiation:
* `input_dtype`
* `ApplyMasksUDF`: Number of masks, back-end (dense, or what kind of sparse, torch, ...) and mask `dtype`
* Possibly `dtype` of result buffers if they are not determined by `input_dtype` -- for example user-provided `dtype`
* Possibly cache size(s) of the CPU it will run on
* With #364: Run on GPU or CPU, which GPU (available RAM -- anything else?)
