### 2020-04-14 LiberTEM/LiberTEM#554 @uellue
DIscussion with @sk1p: Lower and upper bound for signal size does make sense for `kind='sig'` reductions to make sure that the piece of buffer fits the cache.
