### 2020-04-20 #2006 @uellue
@majensen Progress! It just turned into a bigger update, closing some long-standing TODO items in the process: https://github.com/LiberTEM/LiberTEM/pull/706
