### 2020-05-11 #2006 @majensen
@uellue @fedorov (and @alvarolopez ) - checking in. I see Andrey has only the three Functionality items to check off before completion. I know Dieter has been working through a number of improvements. My thought is, if the functionality as it stands (in medias res) is acceptable to Andrey, maybe we can go forward with completion of the review and towards a recommendation for publication. Dieter, you would be free and encouraged to continue development, but with the publication in your pocket.
If the editorial powers have an issue with this plan, they will manifest themselves. What do you guys think?
Mark
