### 2020-05-11 #2006 @uellue
@majensen @fedorov the installation instructions have been fixed, as far as I am aware. We now have some information on reference datasets, too.

The performance bottleneck that we found when addressing the performance tests is mostly fixed. In detail, sparse matrix products with memory-mapped partitions led to large intermediate result buffers that didn't fit the L3 cache. That severely hurt performance of the scipy.sparse library that we use. The same underlying cause applies to a number of other operations. The fix in the back-end is merged LiberTEM/LiberTEM#706, it just has to be used properly in the front-end part of the code, and we have to benchmark it properly. Solving it involved taking on a long-standing item for restructuring the I/O back-end, that's why it took so long.

About related packages, we have it relatively hidden in our acknowledgments more as a hat tip towards other projects, but we could also move it to a more visible location like README.rst if you think it would be a valuable resource for users.

In that sense, I could still work on proper benchmarks, but we can also remove performance claims and wrap up the paper before. :-)
