### 2020-05-11 #2006 @fedorov
> About related packages, we have it relatively hidden in our acknowledgments

What I referred to in my comments were not related packages, but packages similar in functionality. I recall you mention several of those in your paper, but I did not find them in the documentation. I believe it would be helpful for a user who searches for tools implementing certain functionality to understand better how your package is unique relative to other similar packages (if such exist).

> In that sense, I could still work on proper benchmarks, but we can also remove performance claims and wrap up the paper before. :-)

👍
