### 2020-05-18 #2006 @majensen
Thanks guys. Since the associate editor-in-chief will look askance at unchecked items, let me propose that when @fedorov is satisfied with @uellue 's updates to paper, software, or both, that @fedorov can check the remaining two boxes and ping me here.
At that point, I will do a proofing once-over which may result in a minor pull request for wording or reference resolution, then will we call in the EiC.
