### 2020-05-19 #2006 @majensen
@uellue - The proof looks good to me, just two items for you to review:

in paper.md line 140:
* "an adapted programming model" - Do you mean "adaptive" here?

in paper.bib line 5
* "Development of ..." - I note the title at the archive 10.5281/zenodo.2450624 is "Evolution of..."

If you would rectify these to your liking, what we need next is the DOI to an archive containing the repo. The slight issue between the current Zenodo archive and the paper an sich is the author list. JOSS prefers that the author list on the archive match that on the paper. Would this be difficult to adjust?
