### 2020-05-26 #2006 @uellue
Hi @majensen and @fedorov,

Thx for the feedback!

I just opened a pull request in our repository to update the paper. The performance claims in comparison to other projects are removed, since benchmarking would still take some time to implement properly. The LiberTEM reference is updated to 0.5.0. I've asked the co-creators that have since started contributing to LiberTEM if they would like to join as co-authors.

Two co-creators of LiberTEM, Colin Ophus and Jay van Schyndel, had previously declined to be co-authors since the made only small contributions that they felt didn't warrant co-authorship on the paper. Peter Simon and Jaeweon Shin hadn't replied to an invitation to be co-author.

As next steps, I'll wait for response from the new co-authors, update the paper's author list accordingly, merge, and then we can build a new draft here.
