### 2020-05-26 #2006 @fedorov
> Two co-creators of LiberTEM, Colin Ophus and Jay van Schyndel, had previously declined to be co-authors since the made only small contributions that they felt didn't warrant co-authorship on the paper.

It's yours and their call, but IMHO if they made contribution to the code described in the paper, this most definitely justifies their co-authorship on the paper.
