### 2020-05-26 #2006 @uellue
@fedorov yes, I agree that their co-authorship would be justified. We had a discussion about that some time ago and the result was [our authorship policy](https://libertem.github.io/LiberTEM/authorship.html) where everybody who contributed to the repository is invited to be co-author.
