### 2020-05-26 #2006 @fedorov
It is remarkable how organized your community is. Once you mentioned it, it is obvious that Authorship policy is a must to have, but I am sure I could find quite a few of established projects that don't have one.
