### 2020-05-26 #2006 @uellue
Well, we added one when we needed one! :-D The main discussion was actually in the PR for the first draft of this paper: https://github.com/LiberTEM/LiberTEM/pull/460

Main points were that the decision can't be arbitrary, but based on rules, and that we have to remain able to act and fulfill all requirements for authorship of a paper (consent and approval of a draft by all co-authors) even if we have a lot of co-authors in LiberTEM.
