### 2020-06-08 #2006 @uellue
> Great @uellue - can you provide the doi for your archive here- then I will twiddle various knobs and then summon the editors in chief--

@majensen sounds good! This is the DOI for LiberTEM 0.5.0: https://doi.org/10.5281/zenodo.3763313
