### 2020-06-09 #2006 @majensen
@uellue I see the following people not in the paper author list, but listed in the archive

Dey, Shankhadeep;
 Levin, Barnaby D.A.;
 Ophus, Colin;
 Peter, Simon;
 Schyndel van, Jay;
 Shin, Jaeweon;
 Sunku, Sai;

If these folks are not willing to be on the paper, let me know here. I can understand why they would remain on the archive, so I will want to have an EIC weigh in on the protocol in this situation.
