### 2020-06-09 #2006 @uellue
@majensen In the archive we have everybody who contributed a commit or who contributed to LiberTEM in other ways that establish co-authorship of the software. One would have to actively opt out of being listed there since contributing commits to a software automatically establishes co-authorship. Furthermore, being a co-author of the software doesn't mean endorsement or responsibility for the entire content.

However, to be co-author of a paper means responsibility and endorsement for the entire manuscript. That's why we require co-authors to actively opt in per our authorship policy. Some people didn't respond, and some responded that they didn't want to be on the paper. All people with extensive contributions to LiberTEM are also co-authors of the paper.
