### 2020-06-09 #2006 @majensen
@openjournals/joss-eics - Please note that LiberTEM is a significant and ongoing project that has its own policies for author and contributor credit. It also has its own Zenodo archive location, for which 0.5.0 is the latest version and incorporates the comments of the reviewers here.

The author list on this paper is a subset of the author list on the archive, because of the nature of the project's community guidelines. (See [@uellue comment](https://github.com/openjournals/joss-reviews/issues/2006#issuecomment-641097965) and prior.) I hope that there can be a variance allowed in the JOSS policy of requiring the author lists to be the same on both the paper and the archive in this instance. Thanks -- @majensen
