### 2020-06-12 #2006 @labarba
I can see in the comments above the rationale for the different author lists in the Zenodo archive and the JOSS paper. But is there any reason why the Zenodo archive title cannot be made to match the JOSS paper archive?
