### 2020-06-17 #2006 @majensen
Well, the archives for the project LiberTEM have a [history of consistent naming](https://zenodo.org/search?page=1&size=20&q=conceptrecid:%221477847%22&sort=-version&all_versions=True). If I were to give a reason as an advocate for the author, I would say that the JOSS paper is a unique event in its history as a project, associated with the latest version of LiberTEM. Independently of this event, there is an established LiberTEM developer community and user base which has come to expect a certain convention in the naming of the software package that they contribute to and depend on. IMO, JOSS is fortunate in having the opportunity to publish this mature work under its masthead, and the journal might reasonably grant flexibility in this situation.
