### 2020-06-17 #2006 @uellue
@labarba Is this a requirement that there's a 1:1 match with the Zenodo title? It seems that I overlooked it.

The title of the Zenodo deposition is the default one if GitHub integration for releases is activated. We can also call the paper `LiberTEM/LiberTEM: 0.5.0` if you prefer. A more descriptive title would just be more appealing, I figured.

As @majensen wrote, we already have quite a number of releases on Zenodo and it would seem a bit odd to rename it just to match a paper title about it.
