### 2020-06-17 #2006 @labarba
We have a (soft) policy of asking authors to manually edit the metadata of the Zenodo deposit after the release has done its thing, so the titles and author list matches the paper. Of course we don't want you to change the title of the _paper_ to match the repo name. Look, I'm not insisting. That's what we regularly do. If you have your reasons, do as you wish.
