### 2020-06-18 #2006 @uellue
@labarba Thank you for the clarification! If possible, I'd like to leave it as it is.  Since both paper title and archive contain the unique name "LiberTEM", I think that readers will understand that the two are referring to each other even if the titles don't match 100%. :-)
