### 2020-06-20 #2006 @majensen
@labarba Sounds like we are ready to pull the trigger
