### 2020-06-20 #2006 @danielskatz
👋 @uellue - if you can merge https://github.com/LiberTEM/LiberTEM/pull/823, we can finalize the publication
