### 2020-06-20 #2006 @danielskatz
👋 @uellue  - sorry, I missed one - can you also merge https://github.com/LiberTEM/LiberTEM/pull/824
