### 2020-06-22 #2006 @uellue
As a general remark, open writing and open review in combination with GitHub worked really well IMO, and the whole process was a refreshing, pleasant and productive experience. I'll be sure to recommend JOSS for papers like this. 👍
